% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-utils.R
\name{add_vertical_lines_at_g}
\alias{add_vertical_lines_at_g}
\title{Add vertical lines for events (g)}
\usage{
add_vertical_lines_at_g(g, display = TRUE)
}
\arguments{
\item{g}{a vector of events}

\item{display}{whether to display the vertical lines}
}
\value{
a ggplot object
}
\description{
Add vertical lines for events (g)
}
\examples{
g <- c(0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10)
plot_event_study_group(complete_sample, g) + add_vertical_lines_at_g(g)
plot_event_study_group(complete_sample, g) + add_vertical_lines_at_g(g, FALSE)
}
