% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-post-em-params.R
\name{aggregate_ATT_by_latent_type}
\alias{aggregate_ATT_by_latent_type}
\title{Aggregate Average Treatment Effects (LTATTs) Across Groups}
\usage{
aggregate_ATT_by_latent_type(ATT_list, priors_treated)
}
\arguments{
\item{ATT_list}{A list of ATT vectors for each transition pattern.}

\item{priors_treated}{A vector of treated priors for each transition pattern.}
}
\value{
A single vector of aggregated LTATTs.
}
\description{
This function aggregates LTATTs across different transition patterns, weighted by
their respective treated priors.
}
\examples{
ATT_list <- list(c(0.1, 0.2), c(0.3, 0.4))
priors_treated <- c(0.5, 0.5)
aggregated_ATTs <- aggregate_ATT_by_latent_type(ATT_list, priors_treated)

}
