% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical-examples-specs.R
\name{apply_opts_from_empirical_spec}
\alias{apply_opts_from_empirical_spec}
\title{Apply Options from Empirical Specification}
\usage{
apply_opts_from_empirical_spec(opts, empirical_spec)
}
\arguments{
\item{opts}{A list of options to update. If \code{NULL}, a new list is
created.}

\item{empirical_spec}{A list or object containing option fields such as
\code{se_method}.}
}
\value{
A list of updated options, including fields inherited from
\code{empirical_spec}.
}
\description{
Updates a list of options by inheriting relevant settings (e.g., standard
error method) from an empirical specification object.
}
\examples{
empirical_spec <- list(opts_se_method = "boot_multiplier")
apply_opts_from_empirical_spec(list(), empirical_spec)

}
