% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition.R
\name{apply_transition_matrices}
\alias{apply_transition_matrices}
\title{Apply Transition Matrices to a PMF}
\usage{
apply_transition_matrices(initial_pmf, transition_matrices)
}
\arguments{
\item{initial_pmf}{A numeric vector representing the initial PMF.}

\item{transition_matrices}{A list of matrices, where each matrix represents a transition probability matrix.}
}
\value{
A numeric vector representing the resulting PMF after applying all the transition matrices.
}
\description{
This function computes the resulting PMF after sequentially applying a list of transition matrices
to an initial PMF.
}
\examples{
# Example initial PMF
initial_pmf <- c(0.5, 0.3, 0.2)

# Example transition matrices
transition_matrices <- list(
  matrix(c(0.8, 0.1, 0.1, 0.2, 0.6, 0.2, 0.3, 0.3, 0.4), nrow = 3, byrow = TRUE),
  matrix(c(0.7, 0.2, 0.1, 0.2, 0.5, 0.3, 0.4, 0.4, 0.2), nrow = 3, byrow = TRUE)
)

# Compute the resulting PMF
resulting_pmf <- apply_transition_matrices(initial_pmf, transition_matrices)
print(resulting_pmf)

}
