% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-utils.R
\name{arrange_plots}
\alias{arrange_plots}
\title{Arrange Multiple ggplot2 Plots Horizontally}
\usage{
arrange_plots(...)
}
\arguments{
\item{...}{One or more ggplot2 plot objects to be arranged.}
}
\value{
A combined ggplot object with the input plots arranged side by side in a single row
and a common legend at the bottom.
}
\description{
Arranges multiple ggplot2 plot objects side by side in a single row within a single plot area,
ensuring they share a common legend at the bottom. This function is useful for comparing related
visualizations or displaying complementary information across several plots.
}
\examples{
plot1 <- ggplot(mtcars, aes(mpg, disp)) + geom_point(aes(color = factor(cyl))) + theme_minimal()
plot2 <- ggplot(mtcars, aes(hp, wt)) + geom_point(aes(color = factor(cyl))) + theme_minimal()
plot3 <- ggplot(mtcars, aes(qsec, drat)) + geom_point(aes(color = factor(cyl))) + theme_minimal()
combined_plot <- arrange_multiple_plots(plot1, plot2, plot3)
print(combined_plot)

}
