% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{bootstrap}
\alias{bootstrap}
\title{Bootstrap Functions for Transition Models}
\description{
Functions for generating bootstrapped transition models using weighted
bootstrap and multiplier bootstrap methods.
}
