% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-score-matrix.R
\name{cbind_list_of_matrices}
\alias{cbind_list_of_matrices}
\title{Combine List of Matrices Column-wise}
\usage{
cbind_list_of_matrices(l)
}
\arguments{
\item{l}{A list of matrices or a list of lists of matrices.}
}
\value{
A combined matrix with elements bound column-wise.
}
\description{
This function takes a list of matrices or a list of lists of matrices and binds them column-wise.
If the input is a list of matrices, it simply applies \code{cbind}. If the input is a list of lists of matrices,
it recursively applies the function to each sublist and then binds them column-wise.
}
\examples{
mat_list <- list(matrix(1:4, 2, 2), matrix(5:8, 2, 2))
cbind_list_of_matrices(mat_list)
}
\keyword{internal}
