% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{check_if_hard_classification}
\alias{check_if_hard_classification}
\title{Check if hard classification is performed}
\usage{
check_if_hard_classification(opts = list())
}
\arguments{
\item{opts}{A list of options for the transition model.}
}
\value{
A boolean indicating whether hard classification is performed; default is FALSE.
}
\description{
Checks if hard classification is performed in the first stage of the transition model.
}
