% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{check_if_simplify}
\alias{check_if_simplify}
\title{Check if returned estimated model contains minimal elements}
\usage{
check_if_simplify(opts = list())
}
\arguments{
\item{opts}{A list of options for the transition model.}
}
\value{
A boolean indicating whether simplified estimate is returned; default is FALSE.
}
\description{
Checks if returned estimated model contains minimal elements after estimate_model
}
