% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-vectorize.R
\name{completely_vectorize_transition_model}
\alias{completely_vectorize_transition_model}
\title{Vectorize Transition Model Components}
\usage{
completely_vectorize_transition_model(model, minimal = TRUE)
}
\arguments{
\item{model}{A list representing the transition model, containing elements such as
priors, priors_treated, probability mass functions (pmfs), and transition matrices.}

\item{minimal}{Logical; if TRUE, applies minimal vectorization by removing the last column
of transition matrices and omitting the last probability in PMFs.}
}
\value{
A numeric vector consisting of vectorized elements of the transition model.
}
\description{
This function extracts and vectorizes key elements from a transition model,
returning them as a vector
}
