% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-post-em-params.R
\name{compute_priors_treated}
\alias{compute_priors_treated}
\title{Compute Priors for Treated Groups}
\usage{
compute_priors_treated(weights, g, unit_weights = NULL)
}
\arguments{
\item{weights}{An N by J matrix of individual weights of group membership.}

\item{g}{A vector indicating the treatment status for each unit.}

\item{unit_weights}{A vector of individual weights for each unit. If NULL, all
units are assumed to have equal weight.}
}
\value{
A vector of mean posterior probabilities for each treated group.
}
\description{
This function calculates the mean posterior probabilities for the treated
units, effectively estimating the priors for the treated groups.
}
\examples{
weights <- matrix(runif(20), nrow = 10, ncol = 2)
g <- sample(0:1, 10, replace = TRUE)
priors_treated <- compute_priors_treated(weights, g)

}
