% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-post-em-params.R
\name{compute_prob_treated}
\alias{compute_prob_treated}
\title{Compute Probability of Treatment Assignment}
\usage{
compute_prob_treated(g, unit_weights = NULL)
}
\arguments{
\item{g}{A vector indicating the treatment status for each unit.}

\item{unit_weights}{A vector of individual weights for each unit. If NULL, all
units are assumed to have equal weight.}
}
\value{
A numeric value representing the probability of treatment assignment.
}
\description{
This function calculates the probability of treatment assignment by
computing the weighted mean of the treatment indicator.
}
\examples{
g <- sample(0:1, 10, replace = TRUE)
prob_treated <- compute_prob_treated(g)

unit_weights <- runif(10)
prob_treated_weighted <- compute_prob_treated(g, unit_weights)

}
