% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{compute_probs_treated_by_type}
\alias{compute_probs_treated_by_type}
\title{Convert Priors for Treated to Probabilities of Being Treated for Each Latent Group}
\usage{
compute_probs_treated_by_type(priors_treated, priors, prob_treated)
}
\arguments{
\item{priors_treated}{A vector of prior probabilities for each group given the unit is treated.}

\item{priors}{A vector of overall prior probabilities for each group.}

\item{prob_treated}{The overall probability of being treated in the sample.}
}
\value{
A vector of probabilities of being treated for each group.
}
\description{
This function computes the probabilities of being treated based on the priors for treated individuals,
overall priors, and the overall probability of being treated. It ensures that the computed probabilities
do not exceed 1.
}
\examples{
priors_treated <- c(0.6, 0.4)
priors <- c(0.5, 0.5)
prob_treated <- 0.8
probs_treated_by_type <- compute_probs_treated_by_type(priors_treated, priors, prob_treated)
}
