% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{concatenate_Ps}
\alias{concatenate_Ps}
\title{Concatenate Transition Matrices Lists}
\usage{
concatenate_Ps(Ps_pre, Ps_post)
}
\arguments{
\item{Ps_pre}{A list of transition matrices before a certain time point or event.}

\item{Ps_post}{A list of transition matrices after a certain time point or event.}
}
\value{
A list of concatenated transition matrices for each transition pattern.
}
\description{
This function concatenates two lists of transition matrices (Ps). It assumes that
both lists have the same length and correspond to the same number of transition patterns.
}
\examples{
Ps_pre <- list(matrix(1:4, nrow = 2), matrix(5:8, nrow = 2))
Ps_post <- list(matrix(9:12, nrow = 2), matrix(13:16, nrow = 2))
Ps <- concatenate_Ps(Ps_pre, Ps_post)

}
