% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical-examples-paths.R
\name{create_directory_if_needed}
\alias{create_directory_if_needed}
\title{Create Directory if Needed}
\usage{
create_directory_if_needed(directory)
}
\arguments{
\item{directory}{A character string specifying the path to the directory.}
}
\value{
None. Creates the directory if it does not exist.
}
\description{
Ensures that a specified directory exists by creating it if it does not
already exist.
}
