% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did.R
\name{did_conditional_parallel_trends}
\alias{did_conditional_parallel_trends}
\title{Difference-in-Differences with Conditional Parallel Trends}
\usage{
did_conditional_parallel_trends(y, g, lags = 1)
}
\arguments{
\item{y}{A numeric matrix of dimensions N by T, representing the outcome variable.}

\item{g}{A numeric vector of length N, where each element is zero if not treated and is equal to \code{T_treated} if treated in that particular period.}

\item{lags}{An integer specifying the number of lags for the conditional states. Default is 1.}
}
\value{
A numeric value representing the difference-in-differences estimate.
}
\description{
This function calculates the difference-in-differences (DiD) estimates using conditional parallel trends.
It uses the \code{get_conditional_state_matrix} function to compute lagged values of the outcome variable \code{y}.
}
\examples{
y <- matrix(c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12), nrow = 3, byrow = TRUE)
g <- c(0, 0, 2)
did_conditional_parallel_trends(y, g, lags = 1)

}
