% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/att.R
\name{difference_by_transitions}
\alias{difference_by_transitions}
\title{Calculate the Difference in Outcomes by Transition Probabilities}
\usage{
difference_by_transitions(pmf, P_treated, P_control, Y_values)
}
\arguments{
\item{pmf}{A probability mass function (vector) representing the distribution
of initial states.}

\item{P_treated}{A transition matrix for the treated group.}

\item{P_control}{A transition matrix for the control group.}

\item{Y_values}{A vector of outcome values.}
}
\value{
A numeric value representing the difference in expected outcomes
between the treated and control groups.
}
\description{
This function computes the difference in expected outcomes between treated and
control groups based on their respective transition probabilities and the
distribution of initial states.
}
\examples{
pmf <- c(0.5, 0.5)
P_treated <- matrix(c(0.7, 0.4, 0.3, 0.6), nrow = 2)
P_control <- matrix(c(0.6, 0.5, 0.4, 0.5), nrow = 2)
Y_values <- c(0, 1)
difference <- difference_by_transitions(pmf, P_treated, P_control, Y_values)

}
