% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{draw_multipliers}
\alias{draw_multipliers}
\title{Generate Rademacher Distributed Multipliers}
\usage{
draw_multipliers(N)
}
\arguments{
\item{N}{An integer representing the number of multipliers to generate.}
}
\value{
A numeric vector of length N with values -1 or 1.
}
\description{
Generates a vector of N random multipliers from the Rademacher distribution,
where each value is either -1 or 1 with equal probability.
}
\keyword{internal}
