% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first-stage-E-step.R
\name{E_step}
\alias{E_step}
\title{Perform the E-Step of an EM Algorithm}
\usage{
E_step(transition_model, data_for_est)
}
\arguments{
\item{transition_model}{A transition model with specified outcomes and number of latent groups.}

\item{data_for_est}{A list containing the state sequences for control and treated units,
including \code{y_indices_control} and \code{y_indices_treated}.}
}
\description{
This function executes the E-Step in an EM algorithm, computing the
posterior probabilities of latent group membership for each unit
based on observed state sequences and current model parameters.
}
