% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot-se.R
\name{elementwise_quantile}
\alias{elementwise_quantile}
\title{Compute Elementwise Quantiles for Nested Structures}
\usage{
elementwise_quantile(values, probs)
}
\arguments{
\item{values}{A list of values for which quantiles are computed.}

\item{probs}{A numeric vector of probabilities between 0 and 1.}
}
\value{
A scalar, vector, matrix, or nested list of quantiles.
}
\description{
Computes quantiles element-wise for scalar, vector, matrix, or nested
list parameters.
}
