% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot-se.R
\name{elementwise_sd}
\alias{elementwise_sd}
\title{Compute Standard Errors for Nested Structures}
\usage{
elementwise_sd(values)
}
\arguments{
\item{values}{A list of values for which standard errors are computed.}
}
\value{
A scalar, vector, matrix, or nested list of standard errors.
}
\description{
Recursively computes standard errors for scalar, vector, matrix, or
nested list parameters.
}
