% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical-examples-data.R
\name{empirical-examples-data}
\alias{empirical-examples-data}
\title{Data Transformation Functions for Empirical Examples}
\description{
Functions for transforming model outputs into data frames suitable for
analysis and visualization in empirical example analyses.
}
