% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical-examples-specs.R
\name{empirical_specs_df_to_empirical_specs}
\alias{empirical_specs_df_to_empirical_specs}
\title{Convert a Data Frame of Empirical Specifications to a List of empirical_spec Objects}
\usage{
empirical_specs_df_to_empirical_specs(empirical_specs_df)
}
\arguments{
\item{empirical_specs_df}{A data frame where each row represents an empirical specification.}
}
\value{
A list of objects of class \code{"empirical_spec"}, each corresponding to a row in the input data frame.
}
\description{
This function takes a data frame of empirical specifications and converts each row into a list,
assigning the class \code{"empirical_spec"} to each resulting list. The output is a list of these
empirical specification objects.
}
