% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical-examples-models-did-conditional-lag.R
\name{estimate_and_save_did_conditional_lag_model}
\alias{estimate_and_save_did_conditional_lag_model}
\title{Estimate and Save DiD Model}
\usage{
estimate_and_save_did_conditional_lag_model(
  empirical_spec,
  event_study_df,
  lag = 2,
  opts = NULL
)
}
\arguments{
\item{empirical_spec}{A list containing the empirical specification details.}

\item{event_study_df}{A data frame containing the event study data.}

\item{lag}{An integer specifying the number of lags to include in the model.}

\item{opts}{A list of options for the model estimation.}
}
\value{
The estimated model object.
}
\description{
Estimates a did model using specified parameters and saves the result to a file.
}
