% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical-examples-models.R
\name{estimate_and_save_model}
\alias{estimate_and_save_model}
\title{Estimate and Save DID Model}
\usage{
estimate_and_save_model(
  model_name,
  J,
  empirical_spec,
  event_study_df,
  opts = NULL
)
}
\arguments{
\item{model_name}{A character string specifying the model name used for saving.}

\item{J}{An integer specifying the number of lags.}

\item{empirical_spec}{A list containing the empirical specification details.}

\item{event_study_df}{A data frame containing the event study data.}

\item{opts}{A list of options for the model estimation.}
}
\value{
The estimated model object.
}
\description{
Estimates a model using specified parameters and saves the result to a file.
}
