% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-post-em-params.R
\name{estimate_post_em_params}
\alias{estimate_post_em_params}
\title{Estimate Post-Treatment Parameters}
\usage{
estimate_post_em_params(
  posteriors,
  data_for_est,
  unit_weights = NULL,
  opts = list()
)
}
\arguments{
\item{posteriors}{A matrix of posterior probabilities where each column
corresponds to a latent group.}

\item{data_for_est}{A list containing processed data for estimation,
including state sequences and indicators.}

\item{unit_weights}{(Optional) A numeric vector of weights for each unit.}

\item{opts}{(Optional) A list of options for the algorithm.}
}
\value{
A list containing estimated post-treatment parameters:
PMFs for states just before treatment, and transition matrices for
treated and control groups.
}
\description{
This function estimates post-treatment parameters for the second stage of the
algorithm, including transition matrices for treated and control groups and
probability mass functions (PMFs) for states just before treatment.
}
