% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate_transition_model}
\alias{estimate_transition_model}
\title{Estimate a Transition Model for Treatment Effects Analysis}
\usage{
estimate_transition_model(
  J,
  event_study_df,
  yname,
  gname,
  idname,
  tname,
  cluster_varname = NULL,
  outcomes_of_interest = NULL,
  opts = list()
)
}
\arguments{
\item{J}{The number of latent transition groups in the model.}

\item{event_study_df}{long panel data}

\item{yname}{name of dependent variable}

\item{gname}{name of group membership variable}

\item{idname}{name of id variable}

\item{tname}{name of time variable}

\item{cluster_varname}{name of cluster variable (optional)}

\item{opts}{A list of options for the algorithm.}
}
\value{
A list containing the estimated transition model, weights, posteriors, weights, and LTATTs.
}
\description{
This function estimates a transition model for treatment effects analysis by
executing both the first and second stages of the algorithm. It first
constructs a transition model based on pretreatment data and then extends it to
include post-treatment periods.
}
