% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate_transition_model_from_wide_matrix}
\alias{estimate_transition_model_from_wide_matrix}
\title{Estimate a Transition Model for Treatment Effects Analysis}
\usage{
estimate_transition_model_from_wide_matrix(
  J,
  y,
  g,
  c = NULL,
  outcomes_of_interest = NULL,
  unit_weights = NULL,
  transition_model_guesses = NULL,
  opts = NULL
)
}
\arguments{
\item{J}{The number of latent groups in the model.}

\item{y}{A matrix or data frame of outcomes for each unit over time.}

\item{g}{A vector indicating the treatment status for each unit.}

\item{c}{A vector indicating the cluster for each unit (optional)}

\item{outcomes_of_interest}{A string or numeric value for the outcome of interest. Default is maximum}

\item{unit_weights}{A numeric vector of weights for each unit, or \code{NULL}.
If \code{NULL}, equal weights are assumed.}

\item{transition_model_guesses}{A list of initial transition model guesses.
If \code{NULL}, initial guesses are generated internally.}

\item{opts}{A list of options for the algorithm.}
}
\value{
A list containing the estimated transition model attached with
weights, unit_weights, posteriors, with Ps_control_empirical and Ps_treated_empirical
}
\description{
This function estimates a transition model for treatment effects analysis by
executing both the first and second stages of the algorithm. It first
constructs a transition model based on pretreatment data and then extends it to
include post-treatment periods.
}
