% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-utils.R
\name{factor_j_with_priors}
\alias{factor_j_with_priors}
\title{Create Factor with Latent Type Labels}
\usage{
factor_j_with_priors(j, priors)
}
\arguments{
\item{j}{A numeric or factor vector of latent type indices (1, 2, 3, ...).}

\item{priors}{A numeric vector of prior probabilities for each latent type.}
}
\value{
A factor with levels labeled like "1 (85.5\%)", "2 (14.5\%)", etc.
}
\description{
Converts a numeric vector of j values to a factor with proportion labels.
}
