% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-utils.R
\name{factor_outcome_with_share}
\alias{factor_outcome_with_share}
\title{Create Factor with Outcome Share Labels}
\usage{
factor_outcome_with_share(outcomes, outcome_names, outcome_shares)
}
\arguments{
\item{outcomes}{A character or factor vector of outcome names.}

\item{outcome_names}{A character vector of all possible outcome names (in order).}

\item{outcome_shares}{A numeric vector of shares/probabilities for each outcome.}
}
\value{
A factor with levels labeled like "employed (70.5\%)", "unemployed (10.2\%)", etc.
}
\description{
Converts a character/factor vector of outcome names to a factor with share labels.
}
