% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-utils.R
\name{fetch_parameter}
\alias{fetch_parameter}
\title{Fetch Parameter Values from a Model}
\usage{
fetch_parameter(
  model,
  parameter_name,
  j = 1,
  current_index = 1,
  past_index = 1
)
}
\arguments{
\item{model}{A model object containing the parameter to be fetched.}

\item{parameter_name}{The name of the parameter to fetch.}

\item{j}{The group index for which the parameter is fetched (default is 1).}

\item{current_index}{The index corresponding to the current period (default is 1).}

\item{past_index}{The index corresponding to the past period (default is 1).}
}
\value{
The fetched parameter value(s) from the model. Returns \code{NULL} if the parameter does not exist.
}
\description{
This function extracts parameter values from a model object. For matrix-based parameters, it retrieves elements using the specified indices.
}
\examples{
model <- list(parameter_name = list(list(matrix(c(1, 2, 3, 4), nrow = 2))))
fetch_parameter(model, "parameter_name", j = 1, current_index = 1, past_index = 2)
}
\keyword{internal}
