% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-generate.R
\name{find_granularity}
\alias{find_granularity}
\title{Determine the Granularity for the PMF Based on the Length of the Vector}
\usage{
find_granularity(K)
}
\arguments{
\item{K}{An integer representing the number of categories in the PMF.}
}
\value{
A numeric value representing the granularity to be used for generating the PMF.
}
\description{
This function calculates the granularity \code{10^{-J}5} such that \code{K * granularity <= 1}.
If K <= 2, use granularity of 0.1. Otherwise, calculate J such that \code{K * 10^{-J}5 <= 1}.
The granularity ensures that the generated PMF vector elements are multiples of this value,
and their sum is exactly 1.
}
\examples{
granularity <- find_granularity(4)
print(granularity)
}
