% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first-stage.R
\name{generate_first_stage_guess}
\alias{generate_first_stage_guess}
\title{Generate an Initial Guess for the First Stage of an Algorithm}
\usage{
generate_first_stage_guess(y_indices_matrix, J, treated_period = 0)
}
\arguments{
\item{y_indices_matrix}{A matrix of state sequences for each unit.}

\item{J}{The number of latent groups in the model.}

\item{treated_period}{(Optional) The period when treatment starts. Defaults to 0,}
}
\value{
An initial transition model guess, with specified outcomes and number of latent groups.
}
\description{
This function generates an initial guess for the first stage of an algorithm
based on the provided state sequences (\code{y_indices_matrix}) and the number
of latent groups (\code{J}). It creates a transition model with specific outcomes and
time periods.
}
\examples{
\dontrun{
y_indices_matrix <- matrix(sample(1:3, 9, replace = TRUE), nrow = 3)
J <- 2
initial_guess <- generate_first_stage_guess(y_indices_matrix, J)
}
}
