% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-generate.R
\name{generate_initial_outcome_dists}
\alias{generate_initial_outcome_dists}
\title{Generate initial outcome distributions}
\usage{
generate_initial_outcome_dists(Y_t, J)
}
\arguments{
\item{Y_t}{An outcome space object}

\item{J}{Number of transition patterns
matrix and initial outcome dists should be bounded below by some positive number.
Default value is FALSE; when TRUE, every element of transition matrices and
initial outcome dists are guaranteed to be strictly positive.}
}
\value{
A J-length list of pmfs for each transition pattern
}
\description{
Generate initial outcome distributions for each transition pattern
}
