% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-models.R
\name{generate_plot_data_df_for_counterfactuals}
\alias{generate_plot_data_df_for_counterfactuals}
\title{Generate Counterfactual Data for Given Outcomes and Groups}
\usage{
generate_plot_data_df_for_counterfactuals(model, ts, diff_by_status = FALSE)
}
\arguments{
\item{ts}{Time periods to be used.}

\item{diff_by_status}{Logical; if \code{TRUE}, compute differences by treatment status.}

\item{Y_names}{Names of the outcomes.}
}
\value{
A combined data frame of counterfactual outcomes for all subgroups and initial outcomes.
}
\description{
Generate Counterfactual Data for Given Outcomes and Groups
}
