% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-generate.R
\name{generate_pmf}
\alias{generate_pmf}
\title{Generate a Random Probability Mass Function (PMF) for an Arbitrary K}
\usage{
generate_pmf(outcomes)
}
\arguments{
\item{outcomes}{A vector specifying the names of states.}
}
\value{
A numeric vector of length \code{K} representing a PMF where each element is a multiple of the chosen granularity,
and the sum of all elements equals 1.
}
\description{
This function generates a random probability mass function (PMF) vector of length \code{length(outcomes)}
where each element is a multiple of a specified granularity that ensures the sum of the vector is exactly 1.
The granularity is automatically determined based on the value of \code{length(outcomes)}.
}
\examples{
pmf <- generate_pmf_from_K(4)
print(pmf)
}
