% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-generate.R
\name{generate_Ps}
\alias{generate_Ps}
\title{Generate transition matrices}
\usage{
generate_Ps(Y_t, J, T_max, early_null_periods = 0)
}
\arguments{
\item{Y_t}{An outcome space object}

\item{J}{Number of transition patterns}

\item{T_max}{Maximum number of time periods in which transition matrices exist}

\item{early_null_periods}{Number of null periods before transition matrix appears; default = 0
matrix and initial outcome dists should be bounded below by some positive number.
Default value is FALSE; when TRUE, every element of transition matrices and
initial outcome dists are guaranteed to be strictly positive.}
}
\value{
A J-length list of (T_max+early_null_periods-1)-length list of transition matrices for each transition pattern and time period
}
\description{
Generate transition matrices for each transition pattern
}
