% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{generate_sample}
\alias{generate_sample}
\title{Generate a sample from a transition model}
\usage{
generate_sample(transition_model, N = 20)
}
\arguments{
\item{transition_model}{A list representing the transition model. It should include the following elements:
\itemize{
\item Y: An outcome space object.
\item priors: A numeric vector representing the prior probabilities for each group.
\item Ps_control: A list of transition matrices, one for each group.
\item initial_outcome_dist: A list of initial outcome distributions, one for each group.
}}

\item{N}{The number of individuals to sample. Default is 20.}
}
\value{
A list with the following elements:
\itemize{
\item y: A matrix of outcomes for all individuals.
\item y_its: A list of matrices of outcomes, one for each group.
\item N_js: A table showing the number of individuals in each group.
\item transition_model: The input transition model.
}
}
\description{
Generate a sample from a transition model
}
