% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{generate_sample_j}
\alias{generate_sample_j}
\title{Generate a sample for a specific group}
\usage{
generate_sample_j(Y, N_j, Ps_j, initial_outcome_dist_j)
}
\arguments{
\item{Y}{An outcome space object.}

\item{N_j}{The number of individuals to sample for this group.}

\item{Ps_j}{T-list of transition matrices for this group.}

\item{initial_outcome_dist_j}{J-length vector of initial outcome distribution for this group.}
}
\value{
A matrix of outcomes for the individuals in this group.
}
\description{
Generate a sample for a specific group
}
