% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-generate.R
\name{generate_transition_model}
\alias{generate_transition_model}
\title{Generate transition model}
\usage{
generate_transition_model(
  outcomes = c(0, 1),
  J = 2,
  T_max = 10,
  treated_period = 10
)
}
\arguments{
\item{outcomes}{A vector of outcomes}

\item{J}{Number of transition patterns}

\item{T_max}{Maximum number of time periods}

\item{treated_period}{The period in which the treatment is applied; zero or
negative value if never treated, which will make the model not include
Ps_treated}
}
\value{
A list containing the parameters for the transition model with J transition patterns:
\itemize{
\item \code{Y}: An outcome space object
\item \code{priors}: A J-vector of prior probabilities for each transition pattern
\item \code{prob_treated}: A scalar representing the probability of being treated
\item \code{priors_treated}: A J-vector of prior probabilities of belonging to each transition pattern for treated units
\item \code{treated_period}: The period in which the treatment is applied
\item \code{Ps_control}: A J-length list of (T_max-1)-length list of transition
matrices for each transition pattern and time period for control units
\item \code{Ps_treated}: A J-length list of (T_max-1)-length list of transition
if treated_period > 0, otherwise NULL
\item \code{Ps_control_empirical}: A J-length list of (T_max-1)-length list of transition
matrices for each transition pattern and time period for control units,
where the transition matrices are generated from the empirical distribution
of the data
\code{Ps_treated_empirical}: A J-length list of (T_max-1)-length list of transition
matrices for each transition pattern and time period for treated units,
where the transition matrices are generated from the empirical distribution
of the data
\item \code{pmfs_initial_control}: A J-length list of pmfs for initial outcomes for each transition pattern for control units
\item \code{pmfs_initial_treated}: A J-length list of pmfs for initial outcomes for each transition pattern for treated units
\item \code{pmfs_pre_treatment_control}: A J-length list of pmfs for outcomes at (treated_period-1)th period for each transition pattern for control units
\item \code{pmfs_pre_treatment_treated}: A J-length list of pmfs for outcomes at (treated_period-1)th period for each transition pattern for treated units
\item \code{p_y1d}: A J-length list of joint pmfs for (Y_1, D) represented by K by 2 matrices (1st column is for D = 0, 2nd column is for D = 1)
\item 
}
}
\description{
Generate a transition model with parameters
}
