% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{generate_Y_t}
\alias{generate_Y_t}
\title{Generate Outcome Space Details}
\usage{
generate_Y_t(outcomes_observed, outcomes_of_interest = NULL)
}
\arguments{
\item{outcomes_observed}{Numeric / string vector of outcomes for each unit.}

\item{outcomes_of_interest}{String vector of outcomes of interest; default is the last element of sort(unique(y_t))}
}
\value{
A list containing 'values', a sorted numeric vector of unique values or quantiles,
and 'type', a character string indicating whether the outcome space is 'discrete' or 'continuous'.
}
\description{
This function takes a vector of outcomes \code{y_t} and optionally, a list of parameters,
to generate a list describing the outcome space details, which include the type ('discrete' or 'continuous')
and a vector of unique values or quantiles.
}
\examples{
y_t_discrete <- c(1, 2, 2, 3)
result_discrete <- generate_Y_t(y_t_discrete)
print(result_discrete$type)  # Should print 'discrete'
print(result_discrete$values)  # Should print c(0, 0, 1)
result_discrete <- generate_Y_t(y_t_discrete, 2)
print(result_discrete$values)  # Should print c(0, 1, 0)
}
