% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{get_bootstrap_counts}
\alias{get_bootstrap_counts}
\title{Retrieve the Number of Bootstrap Samples}
\usage{
get_bootstrap_counts(opts)
}
\arguments{
\item{opts}{An optional list containing algorithm options.}
}
\value{
An integer representing the number of bootstrap samples.
}
\description{
Retrieves the number of bootstrap samples to be computed based on options.
Default value is 200 if not specified.
}
