% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{get_conditional_state_matrix}
\alias{get_conditional_state_matrix}
\title{Get Conditional State Matrix}
\usage{
get_conditional_state_matrix(y, lags = 1)
}
\arguments{
\item{y}{A numeric matrix.}

\item{lags}{An integer specifying the number of lags. Default is 1.}
}
\value{
A matrix of the same dimensions as \code{y} where each element is a string of concatenated values
separated by underscores.
}
\description{
This function returns a matrix where each element is a concatenation of the current and previous \code{lags} elements
from the input matrix \code{y}. If \code{lags = 1}, each element will be a concatenation of the current and previous element.
If \code{lags = 2}, each element will be a concatenation of the current and previous two elements, and so on.
}
\examples{
y <- matrix(1:6, nrow = 2, byrow = TRUE)
get_conditional_state_matrix(y, lags = 1)
# Returns:
#      [,1]   [,2]   [,3]
# [1,] NA    "1_2" "2_3"
# [2,] NA    "4_5" "5_6"

get_conditional_state_matrix(y, lags = 2)
# Returns:
#      [,1]   [,2]   [,3]
# [1,] NA    NA    "1_2_3"
# [2,] NA    NA    "4_5_6"

}
