% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{get_default_grid_length}
\alias{get_default_grid_length}
\title{Get the Default Grid Length}
\usage{
get_default_grid_length(Y_t_all, ratio = 10)
}
\arguments{
\item{Y_t_all}{A numeric vector containing all the observations in the outcome space \( Y_t \).}

\item{ratio}{An optional numeric value to divide the total number of observations. Defaults to 20.}
}
\value{
An integer representing the default grid length.
}
\description{
This function computes the default grid length given all the observations
in the outcome space \( Y_t \). The grid length is calculated as the ceiling
of the ratio of the total number of observations to a specified ratio, which defaults to 20.
}
\examples{
Y_t_all <- c(1, 2, 3, 4, 5, 6)
get_default_grid_length(Y_t_all) # Should return 1
get_default_grid_length(Y_t_all, ratio = 6) # Should return 1

}
