% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/empirical-examples-plots-counterfactuals-utils.R
\name{get_df_att_did_conditional_lag_models}
\alias{get_df_att_did_conditional_lag_models}
\title{Get Data Frame of ATTs for More Lag Models}
\usage{
get_df_att_did_conditional_lag_models(did_conditional_lag_models)
}
\arguments{
\item{did_conditional_lag_models}{A list of more lagged model objects.}
}
\value{
A data frame containing the ATTs for all more lagged models.
}
\description{
This function retrieves the ATTs for a list of more lagged models.
}
\keyword{internal}
