% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/empirical-examples-plots-counterfactuals-utils.R
\name{get_df_counterfactual_did_conditional_lag_model}
\alias{get_df_counterfactual_did_conditional_lag_model}
\title{Get Data Frame of Counterfactuals for More Lag Model}
\usage{
get_df_counterfactual_did_conditional_lag_model(
  did_conditional_lag_model,
  y_numeric,
  g
)
}
\arguments{
\item{did_conditional_lag_model}{The more lagged model object.}

\item{y_numeric}{The numeric outcome data.}

\item{g}{The treatment assignment vector.}
}
\value{
A data frame containing the counterfactuals for the more lagged model.
}
\description{
This function retrieves the counterfactuals for a more lagged model.
}
