% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/empirical-examples-plots-counterfactuals-utils.R
\name{get_df_counterfactual_did_conditional_lag_models}
\alias{get_df_counterfactual_did_conditional_lag_models}
\title{Get Data Frame of Counterfactuals for More Lag Models}
\usage{
get_df_counterfactual_did_conditional_lag_models(
  did_conditional_lag_models,
  y_numeric,
  g
)
}
\arguments{
\item{did_conditional_lag_models}{A list of more lagged model objects.}

\item{y_numeric}{The numeric outcome data.}

\item{g}{The treatment assignment vector.}
}
\value{
A data frame containing the counterfactuals for all more lagged models.
}
\description{
This function retrieves the counterfactuals for a list of more lagged models.
}
