% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/empirical-examples-plots-counterfactuals-utils.R
\name{get_df_counterfactual_did_model}
\alias{get_df_counterfactual_did_model}
\title{Construct Counterfactual Outcomes Using a Difference-in-Differences Model}
\usage{
get_df_counterfactual_did_model(
  did_model,
  y_numeric,
  g,
  type_name_counterfactual_did = "DiD"
)
}
\arguments{
\item{did_model}{A list containing ATT estimates (\code{att}) and treatment group indicators (\code{group}).}

\item{y_numeric}{A numeric vector of observed outcomes.}

\item{g}{A vector of treatment group indicators.}

\item{type_name_counterfactual_did}{A character string for labeling the DiD counterfactual series. Default is \code{"Counterfactual Untreated (DiD)"}.}
}
\value{
A data frame with the following columns:
\describe{
\item{t}{Time period.}
\item{y}{Estimated untreated (counterfactual) outcome.}
\item{type}{Label identifying the counterfactual series.}
}
}
\description{
Computes a counterfactual outcome trajectory based on a DiD model by subtracting estimated ATT values
from the observed treated outcomes. The result estimates untreated outcomes over time.
}
\seealso{
\code{\link[=get_df_outcome_treated]{get_df_outcome_treated()}}
}
