% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical-examples-plots-counterfactuals.R
\name{get_df_counterfactuals}
\alias{get_df_counterfactuals}
\title{Prepare DataFrame for Counterfactual Analysis}
\usage{
get_df_counterfactuals(model, did_model, y, g, display_J = FALSE)
}
\arguments{
\item{model}{The main counterfactual model.}

\item{did_model}{The DiD counterfactual model.}

\item{y}{Outcome variable matrix.}

\item{g}{Group indicator vector.}

\item{display_J}{Logical, whether to display the number of models.}
}
\value{
A DataFrame of counterfactual outcomes.
}
\description{
This function creates a DataFrame for plotting counterfactual outcomes from the provided models.
}
\keyword{internal}
