% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-data.R
\name{get_differences_in_weighted_means_by_g}
\alias{get_differences_in_weighted_means_by_g}
\title{Compute difference in weighted means by g within each t}
\usage{
get_differences_in_weighted_means_by_g(weighted_mean_by_gjt)
}
\arguments{
\item{weighted_mean_by_gjt:}{Dataframe with weighted means by g, j, and t}
}
\value{
Dataframe with columns of j, t, difference in means, and se of the difference
}
\description{
Compute difference in weighted means by g within each t
}
