% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical-examples-plots.R
\name{get_empirical_plot_default_height}
\alias{get_empirical_plot_default_height}
\title{Get Default Height for Empirical Plots}
\usage{
get_empirical_plot_default_height()
}
\value{
A numeric value indicating the default height. Default is 5.
}
\description{
Returns the default height (in inches) used for empirical plots.
}
\examples{
get_empirical_plot_default_height()
}
