% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical-examples-plots.R
\name{get_empirical_plot_default_width}
\alias{get_empirical_plot_default_width}
\title{Get Default Width for Empirical Plots}
\usage{
get_empirical_plot_default_width()
}
\value{
A numeric value indicating the default width. Default is 8.
}
\description{
Returns the default width (in inches) used for empirical plots.
}
\examples{
get_empirical_plot_default_width()
}
