% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.R
\name{get_expectation_operator}
\alias{get_expectation_operator}
\title{Calculate the expectation operator \( h_t \)}
\usage{
get_expectation_operator(Y_t)
}
\arguments{
\item{Y_t}{A list containing the 'values' vector and 'type' string.
'values' is a sorted numeric vector and 'type' indicates whether the
outcome space is 'discrete' or 'continuous'.}
}
\value{
A numeric vector \( h_t \) serving as the expectation operator.
}
\description{
This function computes the expectation operator \( h_t \), a column vector
such that t(f) \%*\% h_t gives the expected value.
}
\examples{
Y_t_discrete <- list(values = c(1, 2, 3), type = 'discrete')
get_expectation_operator(Y_t_discrete)

Y_t_continuous <- list(values = c(0.1, 0.3, 0.5), type = 'continuous')
get_expectation_operator(Y_t_continuous)
}
