% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/att.R
\name{get_expected_future_outcome}
\alias{get_expected_future_outcome}
\title{Calculate Expected Future Outcome}
\usage{
get_expected_future_outcome(pmf, P_future, Y_values)
}
\arguments{
\item{pmf}{A probability mass function (vector) representing the distribution
of initial states.}

\item{P_future}{A transition matrix representing future state transitions.
This can be the difference between treated and control transition matrices
to compute the difference in expected outcomes.}

\item{Y_values}{A vector of outcome values corresponding to each state.}
}
\value{
A numeric value representing the expected future outcome.
}
\description{
This function computes the expected future outcome given an initial distribution
of states (pmf), a future transition matrix (P_future), and outcome values (Y_values).
}
\examples{
pmf <- c(0.5, 0.5)
P_future <- matrix(c(0.1, -0.1, -0.1, 0.1), nrow = 2)
Y_values <- c(0, 1)
expected_outcome <- get_expected_future_outcome(pmf, P_future, Y_values)
print(expected_outcome)

}
