% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/att.R
\name{get_expected_outcome}
\alias{get_expected_outcome}
\title{Calculate Expected Outcome}
\usage{
get_expected_outcome(pmf, Y_values)
}
\arguments{
\item{pmf}{A probability mass function (vector) representing the distribution
of initial states.}

\item{Y_values}{A vector of outcome values.}
}
\value{
A numeric value representing the expected outcome based on the pmf and Y_values.
}
\description{
Computes the expected outcome given a probability mass function (pmf) and
a vector of outcome values (Y_values).
}
\examples{
pmf <- c(0.5, 0.5)
Y_values <- c(0, 1)
expected_outcome <- get_expected_outcome(pmf, Y_values)
print(expected_outcome)

}
