% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{get_expected_outcomes_df}
\alias{get_expected_outcomes_df}
\title{Get Expected Outcomes Data Frame from Transition Model}
\usage{
get_expected_outcomes_df(transition_model, aggregate = FALSE)
}
\arguments{
\item{transition_model}{A list containing elements of the transition model, such as priors,
transition matrices (\code{Ps_treated} and \code{Ps_control}), and outcome values (\code{Y$values}).
The transition model should also include \code{treated_period}, the period in which treatment begins,
and \code{T_max}, the maximum time period considered in the model.}

\item{aggregate}{Logical; if \code{TRUE}, the function aggregates outcomes over all groups (j),
providing a summary outcome for each time period (\code{t}) and treatment status (\code{g}).
If \code{FALSE}, outcomes are returned for each group and time period separately.}
}
\value{
A data frame containing columns \code{j} (group index), \code{t} (time period), \code{y} (expected outcome),
\code{g} (treatment status), and \code{occurance_rate} (the rate of occurrence for each group and treatment status).
The data frame presents outcomes for both control (g=0) and treated (g=\code{treated_period}) groups across
all considered time periods.
}
\description{
Computes expected outcomes for control and treated groups across all time periods
based on the specified transition model. It optionally aggregates these outcomes over
all groups. The function calculates both the immediate and future expected outcomes
using initial distributions and transition probabilities provided in the transition model.
}
\examples{
# Assuming a predefined transition_model
outcomes_df <- get_expected_outcomes_df(transition_model)
# To aggregate outcomes over all groups
aggregated_outcomes_df <- get_expected_outcomes_df(transition_model, aggregate = TRUE)

}
