% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first-stage.R
\name{get_first_stage_max_iter}
\alias{get_first_stage_max_iter}
\title{Retrieve Maximum Iteration Count for the First Stage}
\usage{
get_first_stage_max_iter(opts)
}
\arguments{
\item{opts}{An optional list containing the algorithm's options, specifically
the maximum number of iterations under \verb{$max_iter}.}
}
\value{
The maximum number of iterations to use.
}
\description{
This function retrieves the maximum number of iterations to be used in the
first stage of an algorithm. If not specified in the options, a default
value is returned.
}
\examples{
\dontrun{
opts <- list(max_iter = 200)
max_iter <- get_first_stage_max_iter(opts)
# Default value
max_iter_default <- get_first_stage_max_iter(NULL)
}
}
