% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first-stage.R
\name{get_first_stage_max_iter_short_run}
\alias{get_first_stage_max_iter_short_run}
\title{Retrieve Maximum Iterations for Short Run Stage}
\usage{
get_first_stage_max_iter_short_run(opts)
}
\arguments{
\item{opts}{A list of options, may contain \code{max_iter_short_run}.}
}
\value{
Integer maximum iterations for short run.
}
\description{
Returns the maximum number of EM iterations for the short run stage.
Defaults to 20 if not specified.
}
\keyword{internal}
