% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first-stage.R
\name{get_first_stage_multistart_counts}
\alias{get_first_stage_multistart_counts}
\title{Retrieve Multistart Counts for the First Stage}
\usage{
get_first_stage_multistart_counts(opts, J = 1)
}
\arguments{
\item{opts}{An optional list containing the algorithm's options, specifically
the number of multistarts under \verb{$multistart_counts} or a base count
under \verb{$multistart_counts_base}.}

\item{J}{The number of latent groups. The default multistart count scales
linearly with J (base_count * J).}
}
\value{
The number of multistarts to use.
}
\description{
This function gets the number of multistarts to use in the first stage of an
algorithm. If not specified in the options, a default value proportional to
the number of latent groups (J) is returned.
}
\examples{
\dontrun{
opts <- list(multistart_counts = 20)
multistart_counts <- get_first_stage_multistart_counts(opts, J = 2)
# Default value scales with J: J=2 -> 20, J=3 -> 40, ..., capped at 100
multistart_counts_default <- get_first_stage_multistart_counts(NULL, J = 3)  # Returns 40
}
}
