% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first-stage.R
\name{get_first_stage_short_run_counts}
\alias{get_first_stage_short_run_counts}
\title{Retrieve Short Run Multistart Counts}
\usage{
get_first_stage_short_run_counts(opts, J = 1)
}
\arguments{
\item{opts}{A list of options, may contain \code{multistart_short_run_counts}.}

\item{J}{The number of latent groups.}
}
\value{
Integer number of short run guesses.
}
\description{
Returns the number of initial guesses to use in the short run stage of
two-stage multistart optimization. Defaults to the standard multistart_counts
if not specified.
}
\keyword{internal}
