% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first-stage.R
\name{get_first_stage_tol}
\alias{get_first_stage_tol}
\title{Retrieve Tolerance Level for the First Stage}
\usage{
get_first_stage_tol(opts)
}
\arguments{
\item{opts}{An optional list containing the algorithm's options, specifically
the tolerance level under \verb{$tol}.}
}
\value{
The tolerance level for convergence.
}
\description{
This function gets the tolerance level for convergence in the first stage of
an algorithm. If not specified in the options, a default value is returned.
}
\examples{
\dontrun{
opts <- list(tol = 1e-5)
tol <- get_first_stage_tol(opts)
# Default value
tol_default <- get_first_stage_tol(NULL)
}
}
