% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{get_interval_names}
\alias{get_interval_names}
\title{Generate names for a set of intervals}
\usage{
get_interval_names(interval_vector)
}
\arguments{
\item{interval_vector}{A numeric vector of sorted interval endpoints.}
}
\value{
A character vector of interval names.
}
\description{
Given a vector of interval endpoints, this function returns a character vector of interval names in the form "[a, b)".
}
\examples{
get_interval_names(c(0, 0.1, 0.3, 0.5))
# should return c("[0.00, 0.10)", "[0.10, 0.30)", "[0.30, 0.50]")
}
