% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{get_J}
\alias{get_J}
\title{Get Number of Latent Groups}
\usage{
get_J(transition_model)
}
\arguments{
\item{transition_model}{A transition model object containing priors.}
}
\value{
Integer number of latent groups.
}
\description{
Extracts the number of latent groups (J) from a transition model.
}
\keyword{internal}
