% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{get_j_indexed_fields}
\alias{get_j_indexed_fields}
\title{Get Fields Indexed by Latent Group (J)}
\usage{
get_j_indexed_fields()
}
\value{
A character vector of field names.
}
\description{
Returns the list of TransitionModel field names that are indexed by latent
group J and need reordering when model groups are reordered.
}
\keyword{internal}
