% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical-examples-table-contents.R
\name{get_j_weighted_se}
\alias{get_j_weighted_se}
\title{Calculate weighted standard error of a variable}
\usage{
get_j_weighted_se(varname, unit_df_j)
}
\arguments{
\item{varname}{Character string specifying the variable name}

\item{unit_df_j}{Data frame containing unit-level data with weight_j column}
}
\value{
Numeric weighted standard error of the variable
}
\description{
This function computes the weighted standard error for a variable using
the appropriate formula for weighted data.
}
\examples{
# Calculate weighted standard error of income for group 1
income_se_group1 <- get_j_weighted_se("income", group1_data)
}
