% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-utils.R
\name{get_jtpa_plot_theme}
\alias{get_jtpa_plot_theme}
\title{Get a ggplot theme for JTPA set}
\usage{
get_jtpa_plot_theme(y_lab = NULL)
}
\arguments{
\item{y_lab}{the label for the y-axis}
}
\value{
a list of ggplot2::theme objects
}
\description{
Get a ggplot theme for JTPA set
}
\examples{
get_jtpa_plot_theme(c(0,1,2),c(1,2,3))
}
