% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-utils.R
\name{get_line_colors}
\alias{get_line_colors}
\title{Get Line Colors for Latent Groups}
\usage{
get_line_colors(js)
}
\arguments{
\item{js}{A vector of latent group indicators.}
}
\value{
A character vector of hex color codes.
}
\description{
Returns colors from the ak palette for latent groups.
}
\keyword{internal}
