% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical-examples-plots-counterfactuals.R
\name{get_line_colors_counterfactual}
\alias{get_line_colors_counterfactual}
\title{Generate Line Colors for Counterfactual Types}
\usage{
get_line_colors_counterfactual(types, Js)
}
\arguments{
\item{types}{A character vector of model types.}

\item{Js}{A numeric vector indicating the number of models.}
}
\value{
A named vector of colors for each model type.
}
\description{
This function assigns colors to each type of counterfactual model for plotting purposes.
}
\keyword{internal}
