% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-utils.R
\name{get_line_sizes}
\alias{get_line_sizes}
\title{Get Line Sizes for Treatment Groups}
\usage{
get_line_sizes(gs)
}
\arguments{
\item{gs}{A vector of treatment group indicators.}
}
\value{
A numeric vector of line sizes.
}
\description{
Returns line widths for plotting.
}
\keyword{internal}
