% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-utils.R
\name{get_line_types}
\alias{get_line_types}
\title{Get Line Types for Treatment Groups}
\usage{
get_line_types(gs)
}
\arguments{
\item{gs}{A vector of treatment group indicators.}
}
\value{
A character vector of linetypes.
}
\description{
Returns ggplot2 linetypes based on the number of treatment groups.
}
\keyword{internal}
