% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical-examples-plots-counterfactuals.R
\name{get_line_types_counterfactual}
\alias{get_line_types_counterfactual}
\title{Generate Line Types for Counterfactual Types}
\usage{
get_line_types_counterfactual(types, Js)
}
\arguments{
\item{types}{A character vector of model types.}

\item{Js}{A numeric vector indicating the number of models.}
}
\value{
A named vector of line types for each model type.
}
\description{
This function assigns line types to each type of counterfactual model for plotting.
}
\keyword{internal}
