% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.R
\name{get_midpoints}
\alias{get_midpoints}
\title{Calculate the midpoints between intervals}
\usage{
get_midpoints(intervals)
}
\arguments{
\item{intervals}{Numeric vector containing sorted intervals.}
}
\value{
Returns a numeric vector containing the midpoints between intervals.
}
\description{
Calculate the midpoints between intervals
}
\examples{
get_midpoints(c(0.1, 0.3, 0.5))  # should return c(0.2, 0.4)
}
