% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical-examples-paths.R
\name{get_model_name}
\alias{get_model_name}
\title{Generate Model Name}
\usage{
get_model_name(J, empirical_spec)
}
\arguments{
\item{J}{An integer specifying the number of latent groups.}

\item{empirical_spec}{A list containing the empirical specification details.}
}
\value{
A character string representing the model file path.
}
\description{
Generates a standardized model name using the outcome variable name and
number of latent groups.
}
