% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{get_multistart_bootstrap_counts}
\alias{get_multistart_bootstrap_counts}
\title{Retrieve Multistart Counts for Bootstrap Estimation}
\usage{
get_multistart_bootstrap_counts(opts)
}
\arguments{
\item{opts}{An optional list containing algorithm options.}
}
\value{
An integer representing the number of multistart points for bootstrap.
}
\description{
Retrieves the number of multistart points to use during bootstrap estimation.
This is separate from the main estimation's multistart_counts to allow
fewer starting points in bootstrap (since the original estimate is used as
an initial guess). Default value is 10.
}
