% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first-stage.R
\name{get_multistart_mc_core_counts}
\alias{get_multistart_mc_core_counts}
\title{Retrieve the Number of Cores for Parallel Multistart}
\usage{
get_multistart_mc_core_counts(opts)
}
\arguments{
\item{opts}{An optional list containing algorithm options.}
}
\value{
An integer representing the number of cores.
}
\description{
Retrieves the number of cores to use for parallel multistart.
Defaults to min(10, detectCores() - 1) if not specified.
}
\keyword{internal}
