% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first-stage.R
\name{get_multistart_mc_core_counts_short_run}
\alias{get_multistart_mc_core_counts_short_run}
\title{Retrieve the Number of Cores for Short Run Stage}
\usage{
get_multistart_mc_core_counts_short_run(opts)
}
\arguments{
\item{opts}{An optional list containing algorithm options.}
}
\value{
An integer representing the number of cores for short run.
}
\description{
Retrieves the number of cores to use for the short run stage of two-stage multistart.
Defaults to \code{multistart_mc_core_counts} if \code{multistart_mc_core_counts_short_run} is not specified.
}
\keyword{internal}
