% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{get_occurence_rates}
\alias{get_occurence_rates}
\title{Calculate Occurrence Rate for Control and Treated Groups}
\usage{
get_occurence_rates(p_y1d, priors)
}
\arguments{
\item{p_y1d}{A J-length list of joint pmfs for (Y_1, D) represented by K by 2 matrices (1st column is for D = 0, 2nd column is for D = 1)}

\item{priors}{A vector of overall prior probabilities for each group.}
}
\value{
A 2xJ matrix with occurrence rates for control (row 1) and treated (row 2) groups across all J groups.
}
\description{
Computes the occurrence rates for control and treated groups across all
groups defined by the priors. This function uses the treated and overall
priors along with the overall probability of being treated to calculate
the occurrence rate for both control and treated groups for each group \code{j}.
}
\examples{
p_y1d <- list(matrix(c(0.3, 0.2, 0.4, 0.1), ncol = 2),
              matrix(c(0.2, 0.1, 0.1, 0.6), ncol = 2))
priors <- c(0.5, 0.5)
occurance_rate <- get_occurence_rates(p_y1d, priors)
print(occurance_rate)
}
