% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_option}
\alias{get_option}
\title{Get Option Value from Options List}
\usage{
get_option(opts, key, default)
}
\arguments{
\item{opts}{An optional list containing algorithm options.}

\item{key}{The name of the option to retrieve.}

\item{default}{The default value to return if the option is not found.}
}
\value{
The option value if found, otherwise the default value.
}
\description{
Generic helper function to retrieve an option value from an options list.
Returns the specified default if the option is not found or opts is not a list.
}
\examples{
opts <- list(max_iter = 200, tol = 1e-5)
get_option(opts, "max_iter", 100)  # Returns 200
get_option(opts, "missing_key", 50)  # Returns 50
get_option(NULL, "max_iter", 100)  # Returns 100

}
