% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical-examples-paths.R
\name{get_plot_att_comparison_all_directory}
\alias{get_plot_att_comparison_all_directory}
\title{Get Plot ATT Comparison All Directory}
\usage{
get_plot_att_comparison_all_directory(empirical_spec)
}
\arguments{
\item{empirical_spec}{A list containing the empirical specification details.}
}
\value{
A character string representing the file path for comparison plots.
}
\description{
Generates the file path for saving ATT comparison plots.
}
