% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical-examples-paths.R
\name{get_plot_att_comparison_directory}
\alias{get_plot_att_comparison_directory}
\title{Get Plot ATT Comparison Directory}
\usage{
get_plot_att_comparison_directory(empirical_spec, suffix = "")
}
\arguments{
\item{empirical_spec}{A list containing the empirical specification details.}

\item{suffix}{An optional character string to add a suffix to the filename.
Default is "".}
}
\value{
A character string representing the file path for comparison plots.
}
\description{
Generates the file path for saving ATT comparison plots.
}
