% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical-examples-paths.R
\name{get_plot_att_directory}
\alias{get_plot_att_directory}
\title{Get Plot ATT Directory}
\usage{
get_plot_att_directory(empirical_spec, J)
}
\arguments{
\item{empirical_spec}{A list containing the empirical specification details.}

\item{J}{An integer specifying the number of latent groups.}
}
\value{
A character string representing the file path for ATT plots.
}
\description{
Generates the file path for saving ATT (Average Treatment Effect on the
Treated) plots.
}
