% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical-examples-paths.R
\name{get_plot_averages_all_outcome_directory}
\alias{get_plot_averages_all_outcome_directory}
\title{Get Plot Averages (All Outcomes) Directory}
\usage{
get_plot_averages_all_outcome_directory(empirical_spec, J = NULL, suffix = "")
}
\arguments{
\item{empirical_spec}{A list containing the empirical specification details.}

\item{J}{An optional integer specifying the number of latent groups.
Default is NULL.}

\item{suffix}{An optional character string to add a suffix to the filename.
Default is "".}
}
\value{
A character string representing the file path for average plots.
}
\description{
Generates the file path for saving average plots for all outcome states.
}
