% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical-examples-paths.R
\name{get_plot_counterfactuals_by_pretreatment_outcomes_directory}
\alias{get_plot_counterfactuals_by_pretreatment_outcomes_directory}
\title{Get Plot Counterfactuals by Pretreatment Outcomes Directory}
\usage{
get_plot_counterfactuals_by_pretreatment_outcomes_directory(
  empirical_spec,
  J = NULL,
  suffix = ""
)
}
\arguments{
\item{empirical_spec}{A list containing the empirical specification details.}

\item{J}{An integer specifying the number of latent groups. Default is NULL.}

\item{suffix}{An optional character string to add a suffix to the filename.
Default is "".}
}
\value{
A character string representing the file path for counterfactual
plots.
}
\description{
Generates the file path for saving counterfactual plots stratified by
pretreatment outcomes.
}
