% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical-examples-paths.R
\name{get_plot_transitions_diffs_directory}
\alias{get_plot_transitions_diffs_directory}
\title{Get Plot Transitions Differences Directory}
\usage{
get_plot_transitions_diffs_directory(
  empirical_spec,
  J,
  outcome_before = NULL,
  outcome_after = NULL,
  suffix = ""
)
}
\arguments{
\item{empirical_spec}{A list containing the empirical specification details.}

\item{J}{An integer specifying the number of latent groups.}

\item{outcome_before}{An optional character string for the outcome state
before transition. Default is NULL.}

\item{outcome_after}{An optional character string for the outcome state
after transition. Default is NULL.}

\item{suffix}{An optional character string to add a suffix to the filename.
Default is "".}
}
\value{
A character string representing the file path for transition plots.
}
\description{
Generates the file path for saving transition difference plots.
}
