% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical-examples-plots-counterfactuals.R
\name{get_point_shapes_counterfactual}
\alias{get_point_shapes_counterfactual}
\title{Generate Point Shapes for Counterfactual Types}
\usage{
get_point_shapes_counterfactual(types, Js)
}
\arguments{
\item{types}{A character vector of model types.}

\item{Js}{A numeric vector indicating the number of models.}
}
\value{
A named vector of ggplot2 shape codes for each model type.
}
\description{
Assigns point shapes to each type of counterfactual model for plotting.
}
\keyword{internal}
