% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{get_se_method}
\alias{get_se_method}
\title{Retrieve Standard Error Method Option}
\usage{
get_se_method(opts)
}
\arguments{
\item{opts}{An optional list containing algorithm options.}
}
\value{
A character string indicating the SE method.
}
\description{
Retrieves the standard error method specified in the options.
Returns "bootstrap", "bootstrap_multiplier", or "none".
}
