% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-utils.R
\name{get_shape_types}
\alias{get_shape_types}
\title{Get Shape Types for Treatment Groups}
\usage{
get_shape_types(gs)
}
\arguments{
\item{gs}{A vector of treatment group indicators.}
}
\value{
A numeric vector of shape codes.
}
\description{
Returns ggplot2 point shapes based on the number of treatment groups.
}
\keyword{internal}
