% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-utils.R
\name{get_shape_types_for_unique_data}
\alias{get_shape_types_for_unique_data}
\title{Get Shape Types for Unique Data Points}
\usage{
get_shape_types_for_unique_data(gs)
}
\arguments{
\item{gs}{A vector of group indicators.}
}
\value{
A numeric vector of shape codes.
}
\description{
Returns ggplot2 point shapes for distinguishing many unique groups.
}
\keyword{internal}
