% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first-stage-solver.R
\name{get_solver_max_iter}
\alias{get_solver_max_iter}
\title{Retrieve Maximum Iterations for Solver}
\usage{
get_solver_max_iter(opts)
}
\arguments{
\item{opts}{A list of options.}
}
\value{
Integer maximum iterations for the solver.
}
\description{
Retrieve Maximum Iterations for Solver
}
\keyword{internal}
