% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first-stage-solver.R
\name{get_solver_rel_tol}
\alias{get_solver_rel_tol}
\title{Retrieve Relative Tolerance for Solver}
\usage{
get_solver_rel_tol(opts)
}
\arguments{
\item{opts}{A list of options.}
}
\value{
Numeric relative tolerance for the solver.
}
\description{
Retrieve Relative Tolerance for Solver
}
\keyword{internal}
