% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{get_T_max}
\alias{get_T_max}
\title{Get Maximum Time Period}
\usage{
get_T_max(transition_model)
}
\arguments{
\item{transition_model}{A transition model object containing transition matrices.}
}
\value{
Integer number of time periods.
}
\description{
Extracts the maximum number of time periods (T) from a transition model.
}
\keyword{internal}
