% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical-examples-table-utils.R
\name{get_table_colhead}
\alias{get_table_colhead}
\title{Generate LaTeX column header for group number}
\usage{
get_table_colhead(j)
}
\arguments{
\item{j}{Integer specifying the group index}
}
\value{
Character string with formatted LaTeX for group number.
}
\description{
Generate LaTeX column header for group number
}
\keyword{internal}
