% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical-examples-table-utils.R
\name{get_table_percent}
\alias{get_table_percent}
\title{Generate LaTeX percentage text for group}
\usage{
get_table_percent(pi_j)
}
\arguments{
\item{pi_j}{Numeric proportion for the group (between 0 and 1)}
}
\value{
Character string with formatted LaTeX for group percentage.
}
\description{
Generate LaTeX percentage text for group
}
\keyword{internal}
