% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical-examples-paths.R
\name{get_tname}
\alias{get_tname}
\title{Get Time Variable Name}
\usage{
get_tname(empirical_spec)
}
\arguments{
\item{empirical_spec}{A list or object containing the empirical specification,
including (optionally) the field \code{time_index_varname} representing
the time variable.}
}
\value{
A character string representing the name of the time variable.
}
\description{
Returns the name of the time variable used in the empirical specification.
Defaults to "time_index" if not explicitly provided in the specification.
}
