% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition.R
\name{get_transition_matrix}
\alias{get_transition_matrix}
\title{Compute Transition Matrix Between Two Outcome Spaces}
\usage{
get_transition_matrix(y_matrix, Y_t_now, Y_t_forward)
}
\arguments{
\item{y_matrix}{A matrix of size N by 2 representing the outcome histories.
The ith row represents the outcome history of unit i for now and forward, respectively.}

\item{Y_t_now}{A list containing outcome space details for the 'now' period.
It should contain \code{values}, a numeric vector, and \code{type}, a character string.}

\item{Y_t_forward}{A list containing outcome space details for the 'forward' period.
It should contain \code{values}, a numeric vector, and \code{type}, a character string.}
}
\value{
A matrix where the (i,j)th element represents the probability of
transitioning from the ith element of \code{Y_t_now$values} to the jth element
of \code{Y_t_forward$values}.
}
\description{
This function computes the transition matrix that represents the
probability of transitioning from an outcome in \code{Y_t_now} to
an outcome in \code{Y_t_forward}.
}
\examples{
y_matrix <- matrix(c(1, 2, 3, 2, 1, 3), ncol = 2)
Y_t_now <- list(values = c(1, 2, 3), type = 'discrete')
Y_t_forward <- list(values = c(1, 2, 3), type = 'discrete')
transition_matrix <- get_transition_matrix(y_matrix, Y_t_now, Y_t_forward)
print(transition_matrix)
}
