% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-utils.R
\name{get_transition_plot_theme}
\alias{get_transition_plot_theme}
\title{Get a ggplot theme for ak}
\usage{
get_transition_plot_theme(gs, js, shapes = NULL, y_lab = NULL)
}
\arguments{
\item{gs}{a vector of unique cohorts}

\item{js}{a vector of unique groups}

\item{shapes}{a vector of unique shapes}

\item{y_lab}{the label for the y-axis}
}
\value{
a list of ggplot2::theme objects
}
\description{
Get a ggplot theme for ak
}
\examples{
get_transition_plot_theme(c(0,1,2),c(1,2,3))
}
