% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{get_transition_probability_df}
\alias{get_transition_probability_df}
\title{Summarize Transition Probabilities from Transition Model}
\usage{
get_transition_probability_df(transition_model, aggregate = FALSE)
}
\arguments{
\item{transition_model}{A list containing the transition model elements, including
transition matrices (\code{Ps_treated} and \code{Ps_control}), and a vector of
outcome names (\code{Y$names}). The transition model should ideally include
additional elements like \code{treated_period}, \code{T_max}, and \code{priors} to
facilitate comprehensive analysis.}

\item{aggregate}{Logical; if \code{TRUE}, the function aggregates transition
probabilities over all groups, providing a summary for each outcome
transition. If \code{FALSE}, transition probabilities are returned for each
group separately.}
}
\value{
A data frame containing columns for the past outcome (\code{past_outcome}),
future outcome (\code{future_outcome}), transition probability (\code{probability}),
treatment status (\code{treatment_status}), and optionally the group (\code{j})
if not aggregated. This summarizes the transition probabilities for all
outcome combinations under both treated and control conditions.
}
\description{
Generates a data frame that summarizes the transition probabilities for control
and treated groups for every possible combination of future and past outcomes
specified in the transition model. It provides insights into how likely transitions
between outcomes are under different conditions.
}
\examples{
# Assuming a predefined transition_model with Y$names defined
transition_probs_df <- get_transition_probability_df(transition_model)
# To aggregate transition probabilities over all groups
aggregated_transition_probs_df <- get_transition_probability_df(transition_model, aggregate = TRUE)

}
