% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical-examples-table-across-types.R
\name{get_weighted_t_test}
\alias{get_weighted_t_test}
\title{Perform weighted t-test comparing variable values between two groups}
\usage{
get_weighted_t_test(varname, unit_df, model, j_reference, j_compared)
}
\arguments{
\item{varname}{Character string specifying the variable name}

\item{unit_df}{Data frame containing unit-level data}

\item{model}{List containing model results including weights}

\item{j_reference}{Integer specifying the reference group index}

\item{j_compared}{Integer specifying the comparison group index}
}
\value{
List containing t-statistic, p-value, standard error, and mean difference
}
\description{
This function conducts a weighted t-test to compare the values of a specified variable
between two different groups from the model.
}
\examples{
# Compare age between groups 1 and 2
age_test <- get_weighted_t_test("age", unit_data, model, 1, 2)
}
