% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-data.R
\name{get_weighted_transition_probabilities_by_gjt}
\alias{get_weighted_transition_probabilities_by_gjt}
\title{compute weighted mean and trend by g, j, and t}
\usage{
get_weighted_transition_probabilities_by_gjt(
  y,
  g,
  weights,
  y_value,
  y_value_past,
  difference_by_treatment = FALSE
)
}
\arguments{
\item{y}{N by T wide matrix of outcomes}

\item{g}{N-vector of treatment timing for each unit}

\item{weights}{N by K matrix of weights}

\item{y_value}{A scalar value of y to compute transition probabilities}

\item{y_value_past}{A scalar value of y in the past to compute transition probabilities}

\item{difference_by_treatment:}{Boolean indicating if differences in means by treatment should be computed}
}
\value{
dataframe with columns of g, j, t, weighted mean, se; NaN if no past value of y_value_past observed and se returned NA if J > 1
}
\description{
compute weighted mean and trend by g, j, and t
}
\examples{
y <- matrix(c(1,2,3,4,5,6), nrow = 3, ncol = 2)
g <- c(1,0)
weights <- matrix(c(0.6, 0.4), nrow = 2, ncol = 2)
get_weighted_mean_by_gjt(y, g, weights)
}
