% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition.R
\name{get_y_t_by_Y_t_index}
\alias{get_y_t_by_Y_t_index}
\title{Obtain the index of each element in y_t within Y_t$values}
\usage{
get_y_t_by_Y_t_index(y_t, Y_t)
}
\arguments{
\item{y_t}{Numeric vector representing the outcomes for each unit.}

\item{Y_t}{List containing 'values', a sorted numeric vector, and 'type',
a character string indicating if the outcome space is 'discrete' or 'continuous'.
For continuous measures use quantile indices; first index such that
y_{it} <= Y_t$values is TRUE}
}
\value{
Returns a numeric vector of indices indicating the position of
each element of y_t within Y_t$values.
}
\description{
Obtain the index of each element in y_t within Y_t$values
}
\examples{
y_t <- c(1, 2, 3, 3)
Y_t <- list(values = c(1, 2, 3), type = 'discrete')
get_y_t_by_Y_t_index(y_t, Y_t) # should return c(1, 2, 3, 3)

y_t <- c(0, 0.2, 0.4, 0.5)
Y_t <- list(values = c(0.1, 0.3, 0.5), type = 'continuous')
get_y_t_by_Y_t_index(y_t, Y_t) # should return c(1, 2, 3, 3)
}
