% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{get_Y_t_names}
\alias{get_Y_t_names}
\title{Generate names for Y_t}
\usage{
get_Y_t_names(Y_t)
}
\arguments{
\item{Y_t}{A list containing the type ("continuous" or otherwise) and values for Y_t.
For continuous type, it should also contain the lower bound.}
}
\value{
A character vector or numeric vector, depending on the type in Y_t.
}
\description{
Given a list Y_t that describes the type and values of an outcome,
this function generates names for the outcomes.
The function differentiates between continuous and categorical types.
}
\examples{
get_Y_t_names(list(type = "continuous", lower_bound = 0, values = c(0.1, 0.2)))
# should return c("[0.00, 0.10)", "[0.10, 0.20)")

get_Y_t_names(list(type = "categorical", values = c("A", "B")))
# should return c("A", "B")
}
