% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{hard_classification}
\alias{hard_classification}
\title{Perform hard classification}
\usage{
hard_classification(posteriors)
}
\arguments{
\item{posteriors}{A matrix of posteriors from the first stage of the transition model.}
}
\value{
A matrix of weights for the second stage of the transition model.
}
\description{
Performs hard classification on the posteriors from the first stage of the transition model.
}
