% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-decomposition.R
\name{inflow_model_to_plot_data_df}
\alias{inflow_model_to_plot_data_df}
\title{Convert Inflow Model to Plot Data}
\usage{
inflow_model_to_plot_data_df(inflow_model, pmfs_per_j)
}
\arguments{
\item{inflow_model}{A model object representing the inflow process.}

\item{pmfs_per_j}{A list of per-type PMF vectors, where pmfs_per_j[\link{j}] gives
Pr(Y_{T0}=y|D=1,Z=j) for each outcome y.}
}
\value{
A data frame formatted for plotting inflow data.
}
\description{
This function extracts relevant data from an inflow model and formats it for plotting.
}
