% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-utils.R
\name{integer_breaks}
\alias{integer_breaks}
\title{Get integer breaks for x or y scale}
\usage{
integer_breaks(n = 5, ...)
}
\arguments{
\item{n}{the number of breaks}
}
\value{
a function that returns integer breaks
}
\description{
Get integer breaks for x or y scale
}
