% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{is_discrete}
\alias{is_discrete}
\title{Check if Outcome Space \( y_t \) is Discrete}
\usage{
is_discrete(y_t)
}
\arguments{
\item{y_t}{A numeric vector representing the outcome space \( y_t \).}
}
\value{
A logical value indicating whether \( y_t \) is discrete or not.
}
\description{
This function checks whether the given outcome space \( y_t \) is discrete.
Specifically, it returns TRUE if there are duplicate values in \( y_t \),
otherwise FALSE.
}
\examples{
is_discrete(c(1, 1, 2)) # Should return TRUE
is_discrete(c(1, 2, 3)) # Should return FALSE
}
