% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{is_Y_t_continuous}
\alias{is_Y_t_continuous}
\title{Check if Outcome Space is Continuous}
\usage{
is_Y_t_continuous(Y_t)
}
\arguments{
\item{Y_t}{An outcome space object with a \code{type} field.}
}
\value{
TRUE if continuous, FALSE otherwise.
}
\description{
Check if Outcome Space is Continuous
}
\keyword{internal}
