% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{is_Y_t_discrete}
\alias{is_Y_t_discrete}
\title{Check if Outcome Space is Discrete}
\usage{
is_Y_t_discrete(Y_t)
}
\arguments{
\item{Y_t}{An outcome space object with a \code{type} field.}
}
\value{
TRUE if discrete, FALSE otherwise.
}
\description{
Check if Outcome Space is Discrete
}
\keyword{internal}
