% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical-examples-plots-counterfactuals.R
\docType{data}
\name{LINECOLORS_DID}
\alias{LINECOLORS_DID}
\title{Line Colors for Counterfactual Models}
\format{
An object of class \code{character} of length 1.
}
\usage{
LINECOLORS_DID
}
\description{
These constants define the line colors used for displaying outcomes in plots.
}
\keyword{datasets}
