% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical-examples-models.R
\name{load_model}
\alias{load_model}
\title{Load Model from File}
\usage{
load_model(filename)
}
\arguments{
\item{filename}{A character string specifying the transition to the saved model file.}
}
\value{
The loaded model object if only one object is present in the file. NULL if the model does not exist
}
\description{
Loads a saved model from a specified file. If there are multiple objects in the file, an error is raised.
}
