% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical-examples-models-did-conditional-lag.R
\name{load_or_estimate_did_conditional_lag_model}
\alias{load_or_estimate_did_conditional_lag_model}
\title{Load or estimate a model with more lags, using did package}
\usage{
load_or_estimate_did_conditional_lag_model(
  empirical_spec,
  event_study_df,
  lag = 2,
  opts = NULL,
  force_estimate = FALSE
)
}
\arguments{
\item{empirical_spec}{A list containing the empirical specification details.}

\item{event_study_df}{A data frame containing the event study data.}

\item{lag}{An integer specifying the number of lags to include in the model.}

\item{opts}{A list of options for the model estimation.}

\item{force_estimate}{A logical value. If TRUE, forces estimation even if a model file exists.}
}
\value{
The loaded or estimated model object.
}
\description{
Loads a saved model if it exists, otherwise estimates a new model and saves it.
If \code{force_estimate} is TRUE, a new model will be estimated and saved even if
an existing model file is found.
}
