% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{long_df_to_wide}
\alias{long_df_to_wide}
\title{Change long panel data to wide}
\usage{
long_df_to_wide(data, varname, idname, tname)
}
\arguments{
\item{data}{long panel data}

\item{varname}{name of variable to extract}

\item{idname}{name of id variable}

\item{tname}{name of time variable}
}
\value{
Matrix of variable varname
}
\description{
Change long panel data to wide
}
\examples{
data <- data.frame(id = c(1,1,1,2,2,2), t = c(1,2,3,1,2,3),
y = c(1,2,3,4,5,6), g = c(1,1,1,2,2,2))
data_to_y_and_gs(data, "y", "g", "id", "t")
}
