% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{long_df_to_y_and_g}
\alias{long_df_to_y_and_g}
\title{Extract y matrix and g vector from long panel data}
\usage{
long_df_to_y_and_g(data, yname, gname, idname, tname, cluster_varname = NULL)
}
\arguments{
\item{data}{long panel data}

\item{yname}{name of dependent variable}

\item{gname}{name of group membership variable}

\item{idname}{name of id variable}

\item{tname}{name of time variable}

\item{cluster_varname}{name of cluster variable (optional); default is NULL}
}
\value{
list with y matrix and g vector
}
\description{
Extract y matrix and g vector from long panel data
}
\examples{
data <- data.frame(id = c(1,1,1,2,2,2), t = c(1,2,3,1,2,3),
y = c(1,2,3,4,5,6), g = c(1,1,1,2,2,2))
long_df_to_y_and_g(data, "y", "g", "id", "t")
}
