% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first-stage-M-step-by-j.R
\name{M_step_initial_outcome_dists_j}
\alias{M_step_initial_outcome_dists_j}
\title{Compute Initial Distribution for a Specific Latent Group}
\usage{
M_step_initial_outcome_dists_j(posteriors_j, initial_outcome_indicators)
}
\arguments{
\item{posteriors_j}{A vector of posterior probabilities for a specific latent group.}

\item{initial_outcome_indicators}{A matrix where each row is a one-hot encoded
vector representing the initial state of an individual.}
}
\value{
A normalized vector representing the initial distribution for the specified
latent group.
}
\description{
Given posterior probabilities and initial outcome indicators, this function computes
the initial distribution for a specific latent group. It calculates the weighted sum
of the initial outcome indicators and normalizes them to form a probability distribution.
}
\examples{
\dontrun{
posteriors_j <- runif(2)
initial_outcome_indicators <- matrix(c(1, 0, 0, 1), nrow = 2)
initial_outcome_dists_j <- M_step_initial_outcome_dists_j(posteriors_j, initial_outcome_indicators)
}
}
