% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first-stage-M-step-by-j.R
\name{M_step_P_jt}
\alias{M_step_P_jt}
\title{Compute Transition Matrix for a Specific Period and Latent Group}
\usage{
M_step_P_jt(posteriors_j, transition_indicators_t)
}
\arguments{
\item{posteriors_j}{A vector of posterior probabilities for a specific latent group.}

\item{transition_indicators_t}{Transition indicators for a specific time period.}
}
\value{
A transition matrix for the specified time period and latent group.
}
\description{
This function computes the transition matrix for a specific time period
and latent group, using the provided posterior probabilities and transition indicators.
}
\examples{
\dontrun{
posteriors_j <- matrix(runif(4), ncol = 2)
transition_indicators_t <- matrix(1:4, nrow = 2)
P_jt <- M_step_P_jt(posteriors_j, transition_indicators_t)
}
}
