% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first-stage-M-step-by-j.R
\name{M_step_Ps_j}
\alias{M_step_Ps_j}
\title{Compute Transition Matrices for Each Period for a Specific Latent Group}
\usage{
M_step_Ps_j(posteriors_j, transition_indicators)
}
\arguments{
\item{posteriors_j}{A vector of posterior probabilities for a specific latent group.}

\item{transition_indicators}{A list of transition indicators for each time period.}
}
\value{
A list of transition matrices for each period for the specified latent group.
}
\description{
Given posterior probabilities and transition indicators, this function computes
the transition matrices for each period for a specific latent group.
}
\examples{
\dontrun{
posteriors_j <- runif(2)
transition_indicators <- list(matrix(1:4, nrow = 2), matrix(5:8, nrow = 2))
Ps_j <- M_step_Ps_j(posteriors_j, transition_indicators)
}
}
