% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-decomposition.R
\name{model_to_decomposition_plot_data_df}
\alias{model_to_decomposition_plot_data_df}
\title{Generate Decomposition Plot Data}
\usage{
model_to_decomposition_plot_data_df(model, outcome_of_interest = NULL)
}
\arguments{
\item{model}{A model object.}

\item{outcome_of_interest}{The primary outcome to focus on, defaults to the last outcome.}
}
\value{
A data frame containing decomposition plot data.
}
\description{
This function generates data for decomposition plots.
}
