% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-score-matrix.R
\name{model_to_score_matrix}
\alias{model_to_score_matrix}
\title{Convert Model to Score Matrix}
\usage{
model_to_score_matrix(y, g, model)
}
\arguments{
\item{y}{Response variable.}

\item{g}{Grouping variable.}

\item{model}{A fitted model containing priors and transition matrices.}
}
\value{
A score matrix representing the structured parameter estimates.
}
\description{
This function computes the score matrix from a given model and data. It extracts relevant probability mass functions
and transition matrices from the model, vectorizes them, and combines them into a single score matrix.
}
\keyword{internal}
