% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot-se.R
\name{models_to_quantiles}
\alias{models_to_quantiles}
\title{Compute Quantiles for Parameters in Transition Models}
\usage{
models_to_quantiles(models, probs)
}
\arguments{
\item{models}{A list of transition models from bootstrap.}

\item{probs}{A numeric vector of probabilities between 0 and 1.}
}
\value{
A list of quantiles for each parameter.
}
\description{
Computes quantiles for various parameters in transition models
based on summaries extracted from a list of bootstrapped models.
}
