% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot-se.R
\name{models_to_standard_errors}
\alias{models_to_standard_errors}
\title{Compute Standard Errors for Parameters in Transition Models}
\usage{
models_to_standard_errors(models)
}
\arguments{
\item{models}{A list of transition models from bootstrap.}
}
\value{
A list of standard errors for each parameter.
}
\description{
Computes standard errors for various parameters in transition models
based on summaries extracted from a list of bootstrapped models.
}
