% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{modify_outcomes_of_interest}
\alias{modify_outcomes_of_interest}
\title{Modify Outcomes of Interest in a Model}
\usage{
modify_outcomes_of_interest(model, outcomes_of_interest)
}
\arguments{
\item{model}{A model object that contains outcome variable information.}

\item{outcomes_of_interest}{A vector of outcome names to update in the model.}
}
\value{
A modified model object with updated outcomes.
}
\description{
This function updates the outcome variables in a model object by
generating new outcome values based on the specified outcomes of interest.
}
\examples{
model <- list(Y = list(names = c("outcome1", "outcome2"), values = c(0, 1)))
new_model <- modify_outcomes_of_interest(model, c("outcome1"))

}
