% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{multiply_matrices_in_list}
\alias{multiply_matrices_in_list}
\title{Multiply Matrices in a List}
\usage{
multiply_matrices_in_list(matrix_list)
}
\arguments{
\item{matrix_list}{A list of matrices.}
}
\value{
The product of the matrices in the list.
}
\description{
This function takes a list of matrices and returns the product of these matrices.
It assumes that the matrices are in a sequence that allows for matrix multiplication
(i.e., the number of columns in one matrix equals the number of rows in the next).
}
\examples{
mat1 <- matrix(1:4, nrow = 2, ncol = 2)
mat2 <- matrix(1:4, nrow = 2, ncol = 2)
result <- multiply_matrices_in_list(list(mat1, mat2))

}
