% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-decomposition.R
\name{outflow_model_to_plot_data_df}
\alias{outflow_model_to_plot_data_df}
\title{Convert Outflow Model to Plot Data}
\usage{
outflow_model_to_plot_data_df(outflow_model, pmfs_per_j, inflow_model)
}
\arguments{
\item{outflow_model}{A model object representing the outflow process.}

\item{pmfs_per_j}{A list of per-type PMF vectors, where pmfs_per_j[\link{j}] gives
Pr(Y_{T0}=y|D=1,Z=j) for each outcome y.}

\item{inflow_model}{The corresponding inflow model for cross-referencing outcomes.}
}
\value{
A data frame formatted for plotting outflow data.
}
\description{
This function extracts relevant data from an outflow model and formats it for plotting.
}
