% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-utils.R
\docType{data}
\name{palette}
\alias{palette}
\title{Color palette for ak plots}
\format{
An object of class \code{character} of length 7.
}
\usage{
palette
}
\description{
A colorblind-friendly palette for plot elements.
}
\keyword{internal}
