% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical-examples-models-comparison.R
\name{plot_all_att_comparison}
\alias{plot_all_att_comparison}
\title{Plot Comparison of Average Treatment Effects (ATT)}
\usage{
plot_all_att_comparison(
  empirical_analysis_result,
  param_name = "ATT",
  display_ci = TRUE,
  suffix = ""
)
}
\arguments{
\item{empirical_analysis_result}{An object containing the results of the empirical analysis.}

\item{param_name}{A character string for the parameter name to plot (default is "ATT").}

\item{display_ci}{Logical, whether to display confidence intervals (default is TRUE).}
}
\value{
A ggplot object showing the comparison of ATT estimates.
}
\description{
This function generates a plot comparing ATT implied by different models.
}
