% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical-examples-models-comparison.R
\name{plot_all_counterfactual_comparison}
\alias{plot_all_counterfactual_comparison}
\title{Plot Comparison of Counterfactuals Implied by Models}
\usage{
plot_all_counterfactual_comparison(
  empirical_analysis_result,
  show_did = TRUE,
  show_ours = TRUE
)
}
\arguments{
\item{empirical_analysis_result}{An object containing the results of the empirical analysis.}

\item{show_did}{Logical, whether to show DiD counterfactuals.}

\item{show_ours}{Logical, whether to show our method's counterfactuals.}
}
\value{
A ggplot object showing the comparison.
}
\description{
This function generates a plot comparing treated and counterfactual outcomes.
}
