% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical-examples-plots.R
\name{plot_data_trends_all_outcome}
\alias{plot_data_trends_all_outcome}
\title{Plot Data Trends (All Outcomes)}
\usage{
plot_data_trends_all_outcome(
  model,
  display_ci = TRUE,
  recenter_t = TRUE,
  display_vertical_event_lines = TRUE,
  ylim_min = -0.1,
  ylim_max = 1.1,
  save = TRUE
)
}
\arguments{
\item{model}{A model object containing the event study data.}

\item{display_ci}{Logical. Whether to display confidence intervals. Default
is TRUE.}

\item{recenter_t}{Logical. Whether to recenter time variable t. Default is
TRUE.}

\item{display_vertical_event_lines}{Logical. Whether to display vertical
event lines. Default is TRUE.}

\item{ylim_min}{Minimum value for the y-axis. Default is -0.1.}

\item{ylim_max}{Maximum value for the y-axis. Default is 1.1.}

\item{save}{A logical value. If TRUE (default), saves the plot to file.}
}
\value{
A ggplot object displaying the data trends.
}
\description{
Generates a plot showing data trends for all outcome states over time,
relative to a policy change.
}
