% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-models.R
\name{plot_ltatt}
\alias{plot_ltatt}
\title{Plot treatment effects by transition pattern group}
\usage{
plot_ltatt(
  model,
  display_ci = FALSE,
  display_vertical_event_lines = TRUE,
  recenter_t = TRUE,
  uniform_ci = TRUE
)
}
\arguments{
\item{model}{A TransitionModel object to be plotted}

\item{display_ci}{Whether to display confidence intervals if available (default is FALSE)}

\item{display_vertical_event_lines}{whether to display vertical lines for events}

\item{recenter_t}{whether to recenter the time axis (g is displayed as 0)}

\item{uniform_ci}{Logical; if \code{TRUE} (default), compute uniform (simultaneous)
confidence intervals; if \code{FALSE}, compute pointwise intervals.}
}
\value{
a ggplot object
}
\description{
Plot treatment effects by transition pattern group
}
