% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-models.R
\name{plot_ltatt_by_pretreatment_outcomes}
\alias{plot_ltatt_by_pretreatment_outcomes}
\title{Plot LTATTs by pretreatment outcomes}
\usage{
plot_ltatt_by_pretreatment_outcomes(model, weighted = FALSE)
}
\arguments{
\item{model}{A TransitionModel object to be plotted}

\item{weighted}{A logical value. If TRUE, display LTATT * pre-treatment pmfs. Default is FALSE.}
}
\value{
A ggplot object visualizing LTATTs by pretreatment outcomes
}
\description{
This function plots latent group average treatment effects on treated (LTATTs)
by pretreatment outcomes, displaying treatment effect heterogeneity across
different initial outcomes just before treatment.
}
