% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-data.R
\name{plot_transition_averages}
\alias{plot_transition_averages}
\title{Plot group averages}
\usage{
plot_transition_averages(
  y,
  g,
  weights = NULL,
  difference_by_treatment = TRUE,
  display_ci = TRUE,
  display_vertical_event_lines = TRUE
)
}
\arguments{
\item{y}{An N by T matrix of outcomes; each row is unit outcome}

\item{g}{N-length vector of treatment status; treated if g > 0}

\item{weights}{An N by J matrix of individual weights of group membership; if NULL assumed to be a single group}

\item{difference_by_treatment}{Boolean indicating if differences in means by treatment should be computed}

\item{display_ci}{whether to show confidence intervals (for J = 1)}

\item{display_vertical_event_lines}{whether to display vertical lines for events}
}
\value{
a ggplot object
}
\description{
Plot group averages
}
