% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-models.R
\name{plot_transition_model_transition_probabilities}
\alias{plot_transition_model_transition_probabilities}
\title{Plot Expected Outcomes from Transition Model}
\usage{
plot_transition_model_transition_probabilities(
  model,
  y_name,
  y_past_name,
  aggregate = FALSE,
  display_vertical_event_lines = TRUE,
  recenter_t = TRUE,
  y_lab = NULL,
  display_ci = TRUE,
  uniform_ci = TRUE
)
}
\arguments{
\item{model}{A list containing the transition model elements, including priors, transition matrices
(\code{Ps_treated} and \code{Ps_control}), and outcome values (\code{Y$values}). The transition model should also
include \code{treated_period}, indicating the period in which treatment begins, and \code{T_max},
the maximum time period considered in the model.}

\item{y_name}{A scalar value of y to compute transition probabilities}

\item{y_past_name}{A scalar value of y in the past to compute transition probabilities}

\item{aggregate}{Logical; if \code{TRUE}, the plot aggregates outcomes over all groups (j), showing a summary
outcome for each time period (\code{t}) and treatment status (\code{g}). If \code{FALSE}, outcomes are plotted
for each group and time period separately.}

\item{display_vertical_event_lines}{whether to display vertical lines for events}

\item{recenter_t}{whether to recenter the time axis (g is displayed as 0)}

\item{y_lab}{the label for the y-axis}

\item{display_ci}{Logical; if \code{TRUE} (default), display confidence intervals.}

\item{uniform_ci}{Logical; if \code{TRUE} (default), compute uniform (simultaneous)
confidence intervals; if \code{FALSE}, compute pointwise intervals.}
}
\value{
a ggplot object
}
\description{
Generates a plot of the expected outcomes for control and treated groups across all time periods
as defined in the transition model. This visualization aids in understanding the dynamic effects
of treatment over time and across different groups.
}
