% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-data.R
\name{plot_transition_probabilities}
\alias{plot_transition_probabilities}
\title{Plot group averages}
\usage{
plot_transition_probabilities(
  model,
  y_name_to = NULL,
  y_name_from = NULL,
  pretreatment_only = TRUE,
  display_ci = TRUE,
  uniform_ci = TRUE,
  display_vertical_event_lines = TRUE,
  recenter_t = TRUE
)
}
\arguments{
\item{y_name_to}{A scalar value of y to compute transition probabilities}

\item{y_name_from}{A scalar value of y in the past to compute transition probabilities}

\item{pretreatment_only}{Boolean indicating if only pretreatment periods should be plotted}

\item{display_ci}{whether to show confidence intervals (for J = 1)}

\item{uniform_ci}{A logical; if \code{TRUE}, compute uniform (simultaneous)
confidence intervals; if \code{FALSE}, compute pointwise intervals.}

\item{display_vertical_event_lines}{whether to display vertical lines for events}

\item{recenter_t}{whether to recenter the time axis (g is displayed as 0)}

\item{y}{An N by T matrix of outcomes; each row is unit outcome}

\item{g}{N-length vector of treatment status; treated if g > 0}

\item{weights}{An N by J matrix of individual weights of group membership; if NULL assumed to be a single group}
}
\value{
a ggplot object
}
\description{
Plot group averages
}
