% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-data.R
\name{plot_weighted_averages}
\alias{plot_weighted_averages}
\title{Plot a data frame of weighted averages}
\usage{
plot_weighted_averages(
  plot_data_df,
  difference_by_treatment = FALSE,
  display_ci = FALSE,
  display_vertical_event_lines = TRUE,
  recenter_t = TRUE
)
}
\arguments{
\item{plot_data_df}{a data frame with columns of g, j, t, weighted mean, and weighted trend}

\item{difference_by_treatment}{Boolean indicating if differences in means by treatment should be computed}

\item{display_ci}{whether to show confidence intervals}

\item{display_vertical_event_lines}{whether to display vertical lines for events}

\item{recenter_t}{whether to recenter the time axis (g is displayed as 0)}
}
\value{
a ggplot object
}
\description{
Plot a data frame of weighted averages
}
\keyword{internal}
