% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical-examples-analysis-results.R
\name{print.empirical_analysis_result}
\alias{print.empirical_analysis_result}
\title{Print method for empirical_analysis_result objects}
\usage{
\method{print}{empirical_analysis_result}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"empirical_analysis_result"}.}

\item{...}{Additional arguments (ignored).}
}
\description{
Summarizes an empirical analysis result, including transition model and DiD outputs.
}
