% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical-examples-specs.R
\name{print.empirical_spec}
\alias{print.empirical_spec}
\title{Print method for empirical_spec objects}
\usage{
\method{print}{empirical_spec}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"empirical_spec"}.}

\item{...}{Additional arguments (ignored).}
}
\description{
Prints a human-readable summary of the empirical specification object.
}
