% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{reorder_transition_model}
\alias{reorder_transition_model}
\title{Reorder Transition Model Estimate by Priors}
\usage{
reorder_transition_model(transition_model)
}
\arguments{
\item{transition_model}{A TransitionModel object to be reordered.}
}
\value{
A reordered TransitionModel object.
}
\description{
Reorders the components of a TransitionModel object based on the priors,
aligning all elements of the transition model according to the order of
the priors.
}
