% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical-examples-models.R
\name{save_model}
\alias{save_model}
\title{Save Model to File}
\usage{
save_model(model, model_name, file_name = NULL, slim = TRUE, compress = "xz")
}
\arguments{
\item{model}{The R object to be saved, usually a trained model.}

\item{model_name}{A character string specifying the name to assign to the model in the global environment.}

\item{file_name}{A character string specifying the file name to save the model. Defaults to \code{model_name.RData}.}

\item{slim}{A logical; if TRUE (default), remove unnecessary components from the model
(especially from bootstrap estimates) to reduce file size.}

\item{compress}{A character string specifying the compression type. Default is "xz" for
best compression. Use "gzip" for faster but larger files, or FALSE for no compression.}
}
\value{
None. The function performs a side effect of saving the model to a file.
}
\description{
This function saves an R object (typically a model) to a \code{.RData} file. The function assigns the object
to a variable in the global environment and saves it to the specified file name.
}
