% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical-examples-plots.R
\name{save_plot}
\alias{save_plot}
\title{Save Plot}
\usage{
save_plot(filename, plot_to_be_saved, width = NULL, height = NULL)
}
\arguments{
\item{filename}{A character string specifying the file transition to save the plot.}

\item{plot_to_be_saved}{The ggplot object to be saved.}

\item{width}{Numeric. Width of the saved plot in inches.
Default is get_empirical_plot_default_width().}

\item{height}{Numeric. Height of the saved plot in inches.
Default is get_empirical_plot_default_height().}
}
\value{
None. The plot is saved to the specified file.
}
\description{
Saves a ggplot object to the specified file, creating the directory if needed.
}
