% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-score-matrix.R
\name{score_matrix_to_se_asymptotic}
\alias{score_matrix_to_se_asymptotic}
\title{Compute Asymptotic Standard Errors from Score Matrix}
\usage{
score_matrix_to_se_asymptotic(score_matrix)
}
\arguments{
\item{score_matrix}{A numeric score matrix.}
}
\value{
A vector of standard errors for each parameter.
}
\description{
This function derives standard errors from the asymptotic variance of the score matrix.
}
\keyword{internal}
