% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-score-matrix.R
\name{score_matrix_to_variance_asymptotic}
\alias{score_matrix_to_variance_asymptotic}
\title{Compute Asymptotic Variance from Score Matrix}
\usage{
score_matrix_to_variance_asymptotic(score_matrix)
}
\arguments{
\item{score_matrix}{A numeric score matrix.}
}
\value{
The asymptotic variance matrix.
}
\description{
This function estimates the asymptotic variance of the parameter estimates based on the score matrix.
}
\keyword{internal}
