% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood-score.R
\name{score_pmf_j}
\alias{score_pmf_j}
\title{Compute Score Function for Initial Outcome Distribution}
\usage{
score_pmf_j(y_indices, pmf_j, posteriors_j, minimal = TRUE)
}
\arguments{
\item{y_indices}{Integer vector of length \eqn{N}, where each entry
represents the observed category index (values in \eqn{\{1, \ldots, K\}}).}

\item{pmf_j}{Numeric vector of length \eqn{K}
representing the initial probability distribution of outcomes.}

\item{posteriors_j}{Numeric vector of length \eqn{N}, representing
posterior probabilities for component \eqn{j} for each observation.}

\item{minimal}{A logical value. If TRUE, returns \eqn{N \times (K-1)} matrix.}
}
\value{
A numeric \eqn{N \times (K-\mathrm{minimal})} matrix, where each row corresponds
to an observation and each column represents the score contribution
for each outcome category.
}
\description{
This function calculates the score function for the initial outcome
distribution of a mixture component \eqn{j}.
}
\examples{
y_indices <- c(1, 2, 1, 3)
pmf_j <- c(0.2, 0.5, 0.3)
posteriors_j <- c(0.7, 0.8, 0.6, 0.9)
score_pmf_j(y_indices, pmf_j, posteriors_j)
}
