% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood-score.R
\name{score_priors}
\alias{score_priors}
\title{Compute Score Function for Mixture Component Priors}
\usage{
score_priors(priors, posteriors, minimal = TRUE)
}
\arguments{
\item{priors}{A numeric vector of length \eqn{J}, representing prior probabilities for each of \eqn{J} mixture components.}

\item{posteriors}{A numeric \eqn{N \times J} matrix, where each entry represents
the posterior probability of observation \eqn{i} belonging to component \eqn{j}.}

\item{minimal}{A logical value. If TRUE, returns \eqn{N \times (J-1)} matrices.}
}
\value{
A numeric \eqn{N \times (J-\mathrm{minimal})} matrix, where each row corresponds to an observation, and each column
represents the score function for a given mixture component prior.
}
\description{
This function calculates the score function for the prior probabilities of mixture components
in a mixture model, given sequences of observed outcomes and posterior probabilities.
}
\examples{
N <- 100
J <- 3
priors <- c(0.3, 0.5, 0.2)
posteriors <- matrix(runif(N * J), nrow = N, ncol = J)
posteriors <- posteriors / rowSums(posteriors)  # Normalize rows to sum to 1
y_indices_matrix <- matrix(sample(1:3, N * 4, replace = TRUE), nrow = N)

scores <- score_priors(y_indices_matrix, priors, posteriors)
print(dim(scores))  # Should be (N, J)

}
