% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood-score-conversion.R
\name{score_to_P_j}
\alias{score_to_P_j}
\title{Convert Score List to Time-Series Transition Matrices}
\usage{
score_to_P_j(P_j_score, P_j, weights = NULL, minimal = TRUE)
}
\arguments{
\item{P_j_score}{A list of \eqn{T-1} elements, where each element is a list of
\eqn{K_t} matrices representing score values for transitions at time step \eqn{t}.}

\item{P_j}{A list of \eqn{T-1} elements, where each element is a \eqn{K_t \times K_{t-1}}
transition probability matrix representing the initial transition probabilities at time step \eqn{t}.}

\item{weights}{An optional numeric vector of length \eqn{N}, representing
observation weights. If provided, scores are weighted accordingly.}

\item{minimal}{A logical value. If TRUE, expects \eqn{N \times (K-1)} matrix
for scores.}
}
\value{
A list of \eqn{T-1} transition probability matrices, where each matrix represents
the updated transition probabilities for a specific time step.
}
\description{
This function converts a list of time-step transition score matrices into
a time-series transition probability model, adjusting it based on an initial
set of transition probability matrices.
}
\examples{
set.seed(123)
T_max <- 3
K_t <- 3
K_t_minus_1 <- 3
P_j_score <- lapply(1:(T_max-1), function(x)
  lapply(1:K_t, function(y) matrix(runif(100 * K_t_minus_1), nrow = 100, ncol = K_t_minus_1))
)
P_j <- lapply(1:(T_max-1), function(x) matrix(1/K_t_minus_1, K_t, K_t_minus_1))  # Uniform initial transitions
weights <- runif(100)

updated_P_j <- score_to_P_j(P_j_score, P_j, weights)
print(length(updated_P_j))  # Should be T-1
}
