% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first-stage.R
\name{select_top_candidates}
\alias{select_top_candidates}
\title{Select Top Candidates by Log-Likelihood}
\usage{
select_top_candidates(estimated_models, k, opts = list())
}
\arguments{
\item{estimated_models}{A list of estimated models, each with a \code{transition_model}
element containing a \code{log_likelihood} field.}

\item{k}{The number of top candidates to select.}
}
\value{
A list of the top K models.
}
\description{
Selects the top K models by log-likelihood from a list of estimated models.
}
\keyword{internal}
