% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize-Ps.R
\name{summarize_Ps}
\alias{summarize_Ps}
\title{Summarize a list of transition matrices}
\usage{
summarize_Ps(Ps, y_names = NULL)
}
\arguments{
\item{Ps}{A J-list of (T-1)-list of transition matrices to be summarized.
Each transition matrix is a KxK matrix representing the probabilities of
transitioning from one state (row) to another (column).}

\item{y_names}{(Optional) A character vector of names for the outcome states.
If not provided, default names will be used.}
}
\value{
A data frame object whose columns \code{j} \code{t} \code{from} \code{to} \code{prob} represent
the transition probabilities for each state in each period for each latent class.
}
\description{
This function summarizes a list of transition matrices by converting them
into a more interpretable format.
≠
}
