% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-summary.R
\name{summarize_transition_model}
\alias{summarize_transition_model}
\title{Summary Method for TransitionModel Objects}
\usage{
summarize_transition_model(model, ...)
}
\arguments{
\item{model}{An object of class \code{TransitionModel}.}

\item{...}{Additional arguments passed to or from other methods (not used here).}
}
\value{
A list containing the summary elements of the transition model.
}
\description{
This function generates a summary for objects of class \code{TransitionModel}.
}
