% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-summary.R
\name{summarize_transition_models}
\alias{summarize_transition_models}
\title{Summarize Multiple Transition Models}
\usage{
summarize_transition_models(models)
}
\arguments{
\item{models}{A list of \code{TransitionModel} objects.}
}
\value{
A list of summaries for each model in \code{models}, where each summary is a \code{TransitionModel} object.
}
\description{
This function generates summaries for a list of \code{TransitionModel} objects.
}
