% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{transition_to_log_likelihood}
\alias{transition_to_log_likelihood}
\title{Estimate the log likelihood of a sequence of outcomes}
\usage{
transition_to_log_likelihood(
  y_indices,
  initial_outcome_dists,
  Ps_control,
  priors
)
}
\arguments{
\item{y_indices}{A vector of indices representing the sequence of outcomes.}

\item{initial_outcome_dists}{A J-length list of JOINT initial outcome distributions (Y_1, D) for each group, fixing treatment status D.}

\item{Ps_control}{A T-length list of transition matrices for group j.}

\item{priors}{A J-length vector of prior probabilities for each group.}
}
\value{
The log likelihood of the sequence of outcomes.
}
\description{
This function estimates the full log likelihood of a sequence of outcomes
given a transition model of initial_outcome_dist and Ps_control.
}
\examples{
\dontrun{
y_indices <- c(1, 2, 1)
priors <- c(0.3, 0.7)
transition_model <- generate_transition_model(c(0,1), length(priors), length(y_indices)-1)
initial_outcome_dists <- transition_model$initial_outcome_dists
Ps_control <- transition_model$Ps_control
transition_to_log_likelihood(y_indices, initial_outcome_dists, Ps_control, priors)
}
}
