% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{transitions_to_log_likelihood}
\alias{transitions_to_log_likelihood}
\title{Estimate the log likelihood of sequences of outcomes}
\usage{
transitions_to_log_likelihood(
  y_indices_matrix,
  initial_outcome_dists,
  Ps_control,
  priors,
  unit_weights = NULL
)
}
\arguments{
\item{y_indices_matrix}{N by (T+1) matrix of indices whose ith row represents ith unit's sequence of outcomes.}

\item{initial_outcome_dists}{A J-length list of JOINT initial outcome distributions (Y_1, D) for each group, fixing treatment status D.}

\item{Ps_control}{A J-length list of transition matrices for each group.}

\item{priors}{A J-length vector of prior probabilities for each group.}

\item{unit_weights}{A numeric vector of weights for each unit, or \code{NULL}.
If \code{NULL}, equal weights are assumed.}
}
\value{
The log likelihood of the sequence of outcomes.
}
\description{
This function estimates the full log likelihood of sequences of outcomes
given a transition model of initial_outcome_dist and Ps_control, optionally applying unit weights.
}
\examples{
\dontrun{
y_indices_matrix <- matrix(c(1, 2, 1, 2, 2, 2), nrow = 2)
priors <- c(0.3, 0.7)
transition_model <- generate_transition_model(c(0,1), length(priors), length(y_indices)-1)
initial_outcome_dists <- transition_model$initial_outcome_dists
Ps_control <- transition_model$Ps_control
transitions_to_log_likelihood(y_indices_matrix, initial_outcome_dists, Ps_control, priors)
}
}
