% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{transitions_to_log_likelihoods_j}
\alias{transitions_to_log_likelihoods_j}
\title{Estimate the log likelihoods of sequences of outcomes}
\usage{
transitions_to_log_likelihoods_j(
  y_indices_matrix,
  initial_outcome_dist_j,
  Ps_control_j
)
}
\arguments{
\item{y_indices_matrix}{N by (T+1) matrix of indices whose ith row represents ith unit's sequence of outcomes.}

\item{initial_outcome_dist_j}{A K-length vector of JOINT initial outcome distributions (Y_1, D) for group j, fixing treatment status D}

\item{Ps_control_j}{A T-length list of transition matrices for group j.}
}
\value{
The log likelihood of the sequence of outcomes.
}
\description{
This function estimates the full log likelihood of a sequence of outcomes
given a transition model of initial_outcome_dist and Ps_control
for each group. The likelihood is given by
\deqn{\sum_{j=1}^J \pi_j f_0^j(y_0) \prod_{t=1}^{T-1} f_t^j(y_t | y_{t-1})}
}
\examples{
y_indices_matrix <- matrix(c(1, 2, 1, 2, 2, 2), nrow = 2)
initial_outcome_dist_j <- c(0.5, 0.5)
Ps_control_j <- list(matrix(c(0.5, 0.5, 0.5, 0.5), nrow = 2),
 matrix(c(0.7, 0.5, 0.3, 0.5), nrow = 2))
transitions_to_log_likelihoods_j(y_indices_matrix, initial_outcome_dist_j, Ps_control_j)
}
