% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{transitions_to_posteriors}
\alias{transitions_to_posteriors}
\title{Estimate the posterior probability of  sequence of outcomes}
\usage{
transitions_to_posteriors(y_indices_matrix, initial_outcome_dists, Ps, priors)
}
\arguments{
\item{y_indices_matrix}{N by (T+1) matrix of indices whose ith row represents ith unit's sequence of outcomes.}

\item{initial_outcome_dists}{A J-length list of JOINT initial outcome distributions (Y_1, D) for each group, fixing treatment status D.}

\item{Ps}{A J-length list of transition matrices for each group.}

\item{priors}{A J-length vector of prior probabilities for each group.}
}
\value{
A N by J matrix of posterior probabilities for each individual belonging to each group.
}
\description{
This function estimates the posterior probability of a sequence of outcomes
given a transition model of initial_outcome_dist and Ps_control
for each group.
}
\examples{
\dontrun{
priors <- c(0.3, 0.7)
transition_model <- generate_transition_model(c(0,1), length(priors), length(y_indices)-1)
initial_outcome_dists <- transition_model$initial_outcome_dists
Ps_control <- transition_model$Ps_control
transitions_to_posteriors(y_indices_matrix, initial_outcome_dists, Ps_control, priors)
}
}
