% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{transitions_to_transition_indicators}
\alias{transitions_to_transition_indicators}
\title{Estimate Transition Indicators for Each Unit and Time Period}
\usage{
transitions_to_transition_indicators(
  y_indices_matrix,
  y_values,
  current_period_to_be_fixed = NULL
)
}
\arguments{
\item{y_indices_matrix}{A numeric matrix of N rows and T+1 columns,
where each row represents a sequence of states (as indices) for
an individual unit over time.}

\item{y_values}{A numeric vector representing the possible states.}

\item{current_period_to_be_fixed}{An integer representing the time period to be
fixed for current states; default is NULL}
}
\value{
A list of length T, where each element is a list of N transition matrices
(each of size K by K). Each matrix in the list corresponds to a single
unit's transition from one state to another in a given time period.
}
\description{
This function computes the transition indicators for each unit across
different time periods. It returns a T-length list where each element is an
N-list of K by K matrices. Each matrix represents the transition pattern
of an individual unit in a particular period.
}
\examples{
y_indices_matrix <- matrix(c(1, 2, 2, 1, 2, 3), nrow = 2)
y_values <- 1:3
transitions <- transitions_to_transition_indicators(y_indices_matrix, y_values)
transitions

}
