% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-utils.R
\name{unfactor}
\alias{unfactor}
\title{Get Factor Levels as Character}
\usage{
unfactor(vector)
}
\arguments{
\item{vector}{A factor vector.}
}
\value{
A character vector.
}
\description{
Extracts factor levels as character strings.
}
\keyword{internal}
