% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-unvectorize-elements.R
\name{unvectorize_pmf}
\alias{unvectorize_pmf}
\title{Unvectorize Probability Mass Function (PMF)}
\usage{
unvectorize_pmf(pmf_vector, minimal = TRUE)
}
\arguments{
\item{pmf_vector}{A numeric vector representing a vectorized PMF.}

\item{minimal}{Logical; if TRUE, reconstructs the last probability to ensure sum = 1.}
}
\value{
A numeric vector representing the original PMF.
}
\description{
This function reconstructs a probability mass function from a vectorized form.
If \code{minimal = TRUE}, it ensures the last probability is reconstructed so the sum equals 1.
}
