% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-unvectorize-elements.R
\name{unvectorize_pmfs}
\alias{unvectorize_pmfs}
\title{Unvectorize Multiple PMF Lists}
\usage{
unvectorize_pmfs(vectorized_pmfs, list_length = 1, minimal = TRUE)
}
\arguments{
\item{list_length}{The number of lists to split the PMFs into.}

\item{minimal}{Logical; if TRUE, reconstructs the last probability to ensure sum = 1.}

\item{pmf_list}{A numeric vector representing multiple concatenated vectorized PMFs.}
}
\value{
A list of lists, where each inner list contains the original PMFs.
}
\description{
This function reconstructs a list of probability mass function (PMF) lists from a single vector.
}
