% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-unvectorize-elements.R
\name{unvectorize_transition_matrices}
\alias{unvectorize_transition_matrices}
\title{Unvectorize Transition Matrix}
\usage{
unvectorize_transition_matrices(vectorized_matrices, nrows, minimal = TRUE)

unvectorize_transition_matrices(vectorized_matrices, nrows, minimal = TRUE)
}
\arguments{
\item{vectorized_matrices}{A numeric vector representing multiple row-wise vectorized transition matrices.}

\item{nrows}{A numeric vector indicating the number of rows in each original transition matrix.}

\item{minimal}{Logical; if TRUE, reconstructs the last column to ensure row sums equal 1.}

\item{vectorized_matrix}{A numeric vector representing a row-wise vectorized transition matrix.}

\item{nrow}{The number of rows in the original transition matrix, i.e.,
the number of current states (vs future)}
}
\value{
A numeric matrix representing the original transition matrix.

A list of numeric matrices representing the original transition matrices.
}
\description{
This function reconstructs a transition matrix from its vectorized form.
If \code{minimal = TRUE}, it ensures that the last column is computed so row sums equal 1.

This function reconstructs a list of transition matrices from a single vector.
If \code{minimal = TRUE}, it ensures the last column of each matrix is reconstructed so row sums equal 1.
}
