% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-unvectorize-elements.R
\name{unvectorize_transition_matrices_list}
\alias{unvectorize_transition_matrices_list}
\title{Unvectorize Multiple Transition Matrix Lists}
\usage{
unvectorize_transition_matrices_list(
  vectorized_matrices_list,
  nrows,
  list_length = 1,
  minimal = TRUE
)
}
\arguments{
\item{vectorized_matrices_list}{A numeric vector representing multiple concatenated vectorized transition matrices.}

\item{nrows}{A numeric vector indicating the number of rows in each original transition matrix.}

\item{list_length}{The number of lists to split the transition matrices into.}

\item{minimal}{Logical; if TRUE, reconstructs the last column to ensure row sums equal 1.}
}
\value{
A list of lists, where each inner list contains the original transition matrices.
}
\description{
This function reconstructs a list of transition matrix lists from a single vector.
}
