% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-unvectorize-elements.R
\name{unvectorize_transition_matrix}
\alias{unvectorize_transition_matrix}
\title{Unvectorize Transition Matrix}
\usage{
unvectorize_transition_matrix(vectorized_matrix, nrow, minimal = TRUE)
}
\arguments{
\item{vectorized_matrix}{A numeric vector representing a row-wise vectorized transition matrix.}

\item{nrow}{The number of rows in the original transition matrix, i.e.,
the number of current states (vs future)}

\item{minimal}{Logical; if TRUE, reconstructs the last column to ensure row sums equal 1.}
}
\value{
A numeric matrix representing the original transition matrix.
}
\description{
This function reconstructs a transition matrix from its vectorized form.
If \code{minimal = TRUE}, it ensures that the last column is computed so row sums equal 1.
}
