% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-unvectorize.R
\name{unvectorize_transition_model}
\alias{unvectorize_transition_model}
\title{Unvectorize Transition Model Components}
\usage{
unvectorize_transition_model(vectorized_model, model_structure, minimal = TRUE)
}
\arguments{
\item{vectorized_model}{A list of numeric vectors representing the vectorized transition model components.}

\item{model_structure}{A list representing the original structure of the transition model (to infer correct dimensions).}

\item{minimal}{Logical; if TRUE, reconstructs missing probabilities in PMFs and transition matrices.}
}
\value{
A list representing the original transition model.
}
\description{
This function reconstructs a transition model from its vectorized components.
}
