% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first-stage.R
\name{use_two_stage_multistart}
\alias{use_two_stage_multistart}
\title{Check if Two-Stage Multistart is Enabled}
\usage{
use_two_stage_multistart(opts)
}
\arguments{
\item{opts}{A list of options, may contain \code{two_stage_multistart}.}
}
\value{
Logical indicating whether to use two-stage multistart.
}
\description{
Determines whether to use the two-stage multistart optimization approach.
This is enabled by default when multistart_counts > 1, but can be disabled
by setting \code{two_stage_multistart = FALSE} in opts.
}
\keyword{internal}
