% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-vectorize-elements.R
\name{vectorize_pmf}
\alias{vectorize_pmf}
\title{Vectorize Probability Mass Function (PMF)}
\usage{
vectorize_pmf(pmf, minimal = TRUE)
}
\arguments{
\item{pmf}{A numeric vector representing the PMF.}

\item{minimal}{Logical; if TRUE, removes the last element.}
}
\value{
A numeric vector representing the vectorized PMF.
}
\description{
This function vectorizes a probability mass function by optionally removing the last element.
}
