% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-vectorize-elements.R
\name{vectorize_pmfs}
\alias{vectorize_pmfs}
\title{Vectorize Multiple PMFs}
\usage{
vectorize_pmfs(pmfs, minimal = TRUE)
}
\arguments{
\item{pmfs}{A list of numeric vectors representing PMFs.}

\item{minimal}{Logical; if TRUE, applies minimal vectorization.}
}
\value{
A list of vectorized PMFs.
}
\description{
This function applies \code{vectorize_pmf} to a list of PMFs.
}
