% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-vectorize-elements.R
\name{vectorize_transition_matrices}
\alias{vectorize_transition_matrices}
\title{Vectorize Multiple Transition Matrices}
\usage{
vectorize_transition_matrices(transition_matrices, minimal = TRUE)
}
\arguments{
\item{transition_matrices}{A list of numeric matrices representing transition matrices.}

\item{minimal}{Logical; if TRUE, applies minimal vectorization.}
}
\value{
A list of vectorized transition matrices.
}
\description{
This function applies \code{vectorize_transition_matrix} to a list of transition matrices.
}
