% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-vectorize-elements.R
\name{vectorize_transition_matrices_list}
\alias{vectorize_transition_matrices_list}
\title{Vectorize Multiple Transition Matrix Lists}
\usage{
vectorize_transition_matrices_list(transition_matrices_list, minimal = TRUE)
}
\arguments{
\item{transition_matrices_list}{A list of lists, where each inner list contains transition matrices.}

\item{minimal}{Logical; if TRUE, applies minimal vectorization by removing the last column.}
}
\value{
A numeric vector representing all vectorized transition matrices from the input list.
}
\description{
This function applies \code{vectorize_transition_matrix} to each transition matrix in a list of lists.
It flattens all transition matrices into a single vector.
}
