% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-vectorize-elements.R
\name{vectorize_transition_matrix}
\alias{vectorize_transition_matrix}
\title{Vectorize Transition Matrix}
\usage{
vectorize_transition_matrix(transition_matrix, minimal = TRUE)
}
\arguments{
\item{transition_matrix}{A numeric matrix representing the transition matrix.}

\item{minimal}{Logical; if TRUE, applies minimal vectorization.}
}
\value{
A numeric vector representing the vectorized transition matrix.
}
\description{
This function vectorizes a transition matrix by applying \code{vectorize_pmf} to each row.
}
