% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-data.R
\name{wide_weighted_sample_to_long_df}
\alias{wide_weighted_sample_to_long_df}
\title{convert wide matrix y and g into long dataframe}
\usage{
wide_weighted_sample_to_long_df(y, g, weights)
}
\arguments{
\item{y:}{wide matrix of outcomes}

\item{g:}{vector of treatment timing for each unit}

\item{weights:}{matrix of weights}
}
\value{
long dataframe with columns of id, y, g, t, and weight
}
\description{
convert wide matrix y and g into long dataframe
}
\examples{
y <- matrix(c(1,2,3,4,5,6), nrow = 3, ncol = 2)
g <- c(1,0)
weights <- matrix(c(0.6, 0.4), nrow = 2, ncol = 2)
wide_weighted_sample_to_long_df(y, g, weights)
}
