% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical-examples-table-utils.R
\name{write_latex_table}
\alias{write_latex_table}
\title{Write a LaTeX table to a file}
\usage{
write_latex_table(table, file_path)
}
\arguments{
\item{table}{Character string containing the LaTeX table code}

\item{file_path}{Character string with the file path}
}
\value{
Invisibly returns the table string.
}
\description{
Write a LaTeX table to a file
}
\keyword{internal}
