% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{y_indices_to_initial_outcome_indicators}
\alias{y_indices_to_initial_outcome_indicators}
\title{Create Initial Outcome Indicators for Each Unit}
\usage{
y_indices_to_initial_outcome_indicators(y_indices_matrix, y_values)
}
\arguments{
\item{y_indices_matrix}{A numeric matrix of N rows and T+1 columns, where
each row represents a sequence of states (as indices) for an individual
unit over time. The first column should represent the initial state.}

\item{y_values}{A numeric vector representing the possible states.}
}
\value{
An N by K matrix, where each row is a one-hot encoded representation
of the initial state of the corresponding unit.
}
\description{
This function generates initial outcome indicators for each unit in the
\code{y_indices_matrix}, using a vectorized approach for efficiency. It returns
an N by K matrix, where each row represents the initial state of each unit
as a one-hot encoded vector.
}
\examples{
y_indices_matrix <- matrix(c(1, 2, 2, 3, 1, 2), nrow = 2)
y_values <- 1:3
initial_outcome_indicators <- y_indices_to_initial_outcome_indicators(y_indices_matrix, y_values)
initial_outcome_indicators

}
