% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{y_to_y_indices}
\alias{y_to_y_indices}
\title{Convert a vector of outcomes to a vector of indices}
\usage{
y_to_y_indices(y, y_values)
}
\arguments{
\item{y}{A vector of outcomes.}

\item{y_values}{A vector of possible outcomes.}
}
\value{
A vector of indices.
}
\description{
Convert a vector of outcomes to a vector of indices
}
\examples{
y_to_y_indices(c("a", "b", "c"), c("a", "b", "c"))
y_to_y_indices(c("a", "b", "c"), c("a", "b"))
y_to_y_indices(c("a", "b", "c"), c("a", "b", "c", "d"))
}
